adds set_precision and fixes tests
