--- tests/unit/test_wkt.py
@@ -1,265 +1,0 @@
-#!/usr/bin/env pytest
-
-import platform
-import pytest
-import shapely
-
-from digital_land.datatype.wkt import WktDataType
-from digital_land.log import IssueLog
-
-
-def issue_type(issues):
-    if issues.rows == []:
-        return None
-    issue = issues.rows.pop()
-    assert issues.rows == []
-    return issue["issue-type"]
-
-
-def test_parse_wkt_polygon():
-    wkt = WktDataType()
-    issues = IssueLog()
-    output = wkt.normalise(
-        "Polygon ((177556.8532 38417.5297999997,177554.4128 38418.3279999997,177549.4009 38419.9708999991,177547.6712 38420.5159000009,177548.2506 38422.5782999992,177549.7508 38422.1910999995,177553.3995 38420.8543999996,177555.9324 38419.9307000004,177557.5129 38419.3800000008,177556.8532 38417.5297999997))",  # noqa: E501
-        issues=issues,
-    )
-    assert (
-        output
-        == "MULTIPOLYGON (((-5.118622 50.203865,-5.118614 50.203882,-5.118724 50.203904,-5.118745 50.203907,-5.118752 50.203888,-5.118728 50.203884,-5.118622 50.203865)))"  # noqa: E501
-    )
-
-
-def test_wkt_point_wgs84():
-    wkt = WktDataType()
-    issues = IssueLog()
-
-    # Nelson's colum
-    assert (
-        wkt.normalise("POINT( -0.127972   51.507722 )", issues=issues)
-        == "POINT (-0.127972 51.507722)"
-    )
-    assert issue_type(issues) is None
-
-
-def test_wkt_point_wgs84_south_west():
-    wkt = WktDataType()
-    issues = IssueLog()
-
-    # Scilly Isles
-    assert (
-        wkt.normalise("POINT (-6.322778 49.936111)", issues=issues)
-        == "POINT (-6.322778 49.936111)"
-    )
-    assert issue_type(issues) is None
-
-
-def test_wkt_point_wgs84_north_east():
-    wkt = WktDataType()
-    issues = IssueLog()
-
-    # Berwick-upon-Tweed
-    assert (
-        wkt.normalise("POINT (-2.007 55.771)", issues=issues)
-        == "POINT (-2.007000 55.771000)"
-    )
-    assert issue_type(issues) is None
-
-
-def test_wkt_point_wgs84_flipped():
-    wkt = WktDataType()
-    issues = IssueLog()
-
-    # Nelson's colum
-    assert (
-        wkt.normalise("POINT (51.507722 -0.127972)", issues=issues)
-        == "POINT (-0.127972 51.507722)"
-    )
-    assert issue_type(issues) == "WGS84 flipped"
-
-
-def test_wkt_point_wgs84_out_of_range():
-    wkt = WktDataType()
-    issues = IssueLog()
-
-    assert wkt.normalise("POINT (0.0 0.0)", issues=issues) == ""
-    assert issue_type(issues) == "WGS84 out of bounds of England"
-
-    assert wkt.normalise("POINT (0.0 48.1)", issues=issues) == ""
-    assert issue_type(issues) == "WGS84 out of bounds of England"
-
-
-def test_wkt_point_northings_eastings():
-    wkt = WktDataType()
-    issues = IssueLog()
-
-    # Nelson's column TQ 30015 80415
-    assert (
-        wkt.normalise("POINT (530015 180415)", issues=issues)
-        == "POINT (-0.127960 51.507718)"
-    )
-    assert issue_type(issues) == "OSGB"
-
-
-def test_parse_wkt_ogbd_to_wgs84_accuracy():
-    wkt = WktDataType()
-    issues = IssueLog()
-    output = wkt.normalise(
-        "MULTIPOLYGON (((177556.8532 38417.5297999997,177554.4128 38418.3279999997,177549.4009 38419.9708999991,177547.6712 38420.5159000009,177548.2506 38422.5782999992,177549.7508 38422.1910999995,177553.3995 38420.8543999996,177555.9324 38419.9307000004,177557.5129 38419.3800000008,177556.8532 38417.5297999997)))",  # noqa: E501
-        issues=issues,
-    )
-    assert (
-        output
-        == "MULTIPOLYGON (((-5.118622 50.203865,-5.118614 50.203882,-5.118724 50.203904,-5.118745 50.203907,-5.118752 50.203888,-5.118728 50.203884,-5.118622 50.203865)))"  # noqa: E501
-    )
-
-
-def test_wkt_point_flipped_northings_eastings():
-    wkt = WktDataType()
-    issues = IssueLog()
-
-    # Nelson's column TQ 30015 80415
-    assert (
-        wkt.normalise("POINT (180415 530015)", issues=issues)
-        == "POINT (-0.127960 51.507718)"
-    )
-    assert issue_type(issues) == "OSGB flipped"
-
-
-def test_wkt_point_mercator():
-    wkt = WktDataType()
-    issues = IssueLog()
-
-    # Nelson's Column
-    # https://epsg.io/map#srs=3857&x=-14245.780102&y=6711600.069496&z=17&layer=streets
-    assert (
-        wkt.normalise("POINT (-14245.780102 6711600.069496)", issues=issues)
-        == "POINT (-0.127972 51.507722)"
-    )
-    assert issue_type(issues) == "Mercator"
-
-
-def test_wkt_point_mercator_flipped():
-    wkt = WktDataType()
-    issues = IssueLog()
-
-    # Nelson's Column
-    # https://epsg.io/map#srs=3857&x=-14245.780102&y=6711600.069496&z=17&layer=streets
-    assert (
-        wkt.normalise("POINT (6711600.069496 -14245.780102)", issues=issues)
-        == "POINT (-0.127972 51.507722)"
-    )
-
-
-@pytest.mark.skipif(
-    platform.system() == "Darwin", reason="Test causes shapely/geos to segfault on OSX"
-)
-def test_wkt_point_missing_values():
-    wkt = WktDataType()
-    issues = IssueLog()
-
-    assert wkt.normalise("POINT", issues=issues) == ""
-    assert issue_type(issues) == "invalid WKT"
-
-    assert wkt.normalise("POINT ()", issues=issues) == ""
-    assert issue_type(issues) == "invalid WKT"
-
-    assert wkt.normalise("POINT (-0.127972 )", issues=issues) == ""
-    assert issue_type(issues) == "invalid WKT"
-
-
-def test_wkt_point_out_of_range_values():
-    wkt = WktDataType()
-    issues = IssueLog()
-
-    assert wkt.normalise("POINT (1000 100000000)", issues=issues) == ""
-    assert issue_type(issues) == "invalid coordinates"
-
-    assert wkt.normalise("POINT (100000000 10000)", issues=issues) == ""
-    assert issue_type(issues) == "invalid coordinates"
-
-
-def test_wkt_multipolygon_wgs84():
-    wkt = WktDataType()
-    issues = IssueLog()
-
-    value = "MULTIPOLYGON (((-0.1434494279 51.46626361,-0.1434646353 51.46627914,-0.143515539 51.4663375,-0.1435648475 51.4663926,-0.1435988703 51.46643054,-0.1436227923 51.46646195,-0.1436840978 51.46644134,-0.1436913831 51.4664392,-0.1437519691 51.46641858,-0.1437548832 51.46641773,-0.1436953554 51.46634835,-0.1435837312 51.46621808,-0.1435209507 51.46623957,-0.1434494279 51.46626361)))"  # noqa: E501
-    expected = "MULTIPOLYGON (((-0.143449 51.466264,-0.143623 51.466462,-0.143755 51.466418,-0.143584 51.466218,-0.143449 51.466264)))"
-    assert wkt.normalise(value, issues=issues) == expected
-    assert issue_type(issues) is None
-
-
-def test_wkt_multipolygon_flipped_northings_and_eastings():
-    wkt = WktDataType()
-    issues = IssueLog()
-    value = "MULTIPOLYGON (((203500.0 494297.28,203499.8 494297.07,203495.1 494292.05,203491.2 494287.55,203487.2 494284.05,203482.45 494280.05,203478.4 494276.3,203479.85 494274.9,203486.95 494265.96,203500.0 494249.55,203503.6 494244.7,203514.5 494230.45,203532.6 494206.8,203554.0 494178.8,203566.2 494162.9,203601.8 494116.4,203626.0 494136.2,203628.4 494138.7,203628.9 494141.1,203626.04 494151.35,203639.0 494160.6,203645.3 494165.9,203650.7 494170.2,203651.8 494170.9,203659.0 494175.7,203670.3 494183.0,203674.8 494186.0,203677.7 494188.6,203619.4 494261.3,203612.3 494270.3,203607.2 494276.8,203602.7 494282.9,203590.9 494299.4,203578.6 494316.9,203578.36 494317.24,203562.9 494338.9,203550.8 494332.5,203546.6 494330.4,203540.7 494326.7,203533.8 494322.3,203522.1 494315.4,203519.1 494313.4,203516.8 494311.9,203514.3 494309.8,203500.0 494297.28)))"  # noqa: E501
-    expected = "MULTIPOLYGON (((-0.636234 51.722310,-0.635995 51.722479,-0.635742 51.722723,-0.635614 51.722868,-0.636501 51.723277,-0.637759 51.723926,-0.637797 51.723900,-0.638032 51.723686,-0.638174 51.723582,-0.638312 51.723467,-0.638459 51.723495,-0.638494 51.723491,-0.638824 51.723255,-0.636925 51.722318,-0.636543 51.722119,-0.636377 51.722233,-0.636234 51.722310)))"  # noqa: E501
-    normalised = wkt.normalise(value, issues=issues)
-    assert normalised == expected
-    assert issue_type(issues) == "OSGB flipped"
-
-
-def test_wkt_multipolygon_mercator():
-    wkt = WktDataType()
-    issues = IssueLog()
-    value = "MULTIPOLYGON (((-7946.4687 6701859.138,-7925.9829 6701856.258,-7926.5076 6701852.447,-7946.5692 6701855.314,-7946.4687 6701859.138)))"  # noqa: E501
-    expected = "MULTIPOLYGON (((-0.071384 51.453226,-0.071385 51.453205,-0.071205 51.453189,-0.071200 51.453210,-0.071384 51.453226)))"  # noqa: E501
-    assert wkt.normalise(value, issues=issues) == expected
-    assert issue_type(issues) == "Mercator"
-
-
-def test_wkt_multipolygon_wgs84_duplicate_polygon():
-    wkt = WktDataType()
-    issues = IssueLog()
-    # Buckinghamshire tree preservation zone
-    value = "MULTIPOLYGON (((-0.111483 51.472552,-0.111481 51.472541,-0.111373 51.472542,-0.111372 51.472551,-0.111483 51.472552)),((-0.111483 51.472541,-0.111481 51.472541,-0.111481 51.472541,-0.111483 51.472541)))"  # noqa: E501
-    expected = "MULTIPOLYGON (((-0.111483 51.472552,-0.111481 51.472541,-0.111373 51.472542,-0.111372 51.472551,-0.111483 51.472552)))"
-    assert wkt.normalise(value, issues=issues) == expected
-
-
-def test_wkt_invalid_multipolygon_too_few_points():
-    wkt = WktDataType()
-    issues = IssueLog()
-    # 465879 "Teesmouth and Cleveland Coast"
-    # Too few points in geometry component[-1.240621 54.600571]
-    value = "MULTIPOLYGON(((-1.240621 54.600571,-1.240618 54.600566,-1.24038 54.600608,-1.240094 54.600681,-1.239689 54.6008,-1.239365 54.60088,-1.239192 54.600903,-1.239082 54.600905,-1.23893 54.600894,-1.238186 54.600798,-1.237731 54.600706,-1.237607 54.600704,-1.237373 54.600645,-1.237228 54.600627,-1.237181 54.600612,-1.237163 54.60059,-1.237159 54.6005,-1.237124 54.60044,-1.237062 54.60039,-1.236918 54.600298,-1.236722 54.600197,-1.236307 54.600009,-1.236144 54.599908,-1.23606 54.599837,-1.235639 54.599828,-1.235434 54.599864,-1.235202 54.599854,-1.235063 54.599822,-1.234897 54.599755,-1.234799 54.599734,-1.234647 54.59972,-1.234565 54.59969,-1.234484 54.599638,-1.234447 54.599608,-1.234405 54.599557,-1.234306 54.599471,-1.234267 54.599451,-1.234022 54.599388,-1.233858 54.599361,-1.233446 54.599375,-1.233442 54.599417,-1.233678 54.601451,-1.233948 54.603642,-1.236817 54.603639,-1.246748 54.603569,-1.250048 54.601986,-1.250198 54.601952,-1.249817 54.601682,-1.248889 54.601045,-1.248357 54.600664,-1.247944 54.600349,-1.246529 54.600752,-1.245915 54.599941,-1.245832 54.599909,-1.245807 54.599911,-1.245686 54.599962,-1.245584 54.599994,-1.245013 54.600219,-1.244672 54.600364,-1.244425 54.600494,-1.244167 54.600571,-1.244048 54.600593,-1.244054 54.600605,-1.243709 54.600674,-1.243382 54.60082,-1.242351 54.601124,-1.242329 54.601101,-1.241955 54.600894,-1.241885 54.600845,-1.241862 54.600805,-1.241837 54.600792,-1.241697 54.600757,-1.241631 54.600749,-1.24158 54.600731,-1.241436 54.600726,-1.241406 54.600709,-1.241349 54.600658,-1.241331 54.600653,-1.241279 54.600595,-1.241261 54.600557,-1.241205 54.600525,-1.241081 54.600521,-1.240771 54.600548,-1.240621 54.600571),(-1.240621 54.600571,-1.240621 54.600571,-1.240592 54.600575,-1.240621 54.600571)))"  # noqa: E501
-    expected = "MULTIPOLYGON (((-1.240771 54.600548,-1.240621 54.600571,-1.240618 54.600566,-1.240380 54.600608,-1.240094 54.600681,-1.239689 54.600800,-1.239365 54.600880,-1.239192 54.600903,-1.239082 54.600905,-1.238930 54.600894,-1.238186 54.600798,-1.237731 54.600706,-1.237607 54.600704,-1.237373 54.600645,-1.237228 54.600627,-1.237181 54.600612,-1.237163 54.600590,-1.237159 54.600500,-1.237124 54.600440,-1.237062 54.600390,-1.236918 54.600298,-1.236722 54.600197,-1.236307 54.600009,-1.236144 54.599908,-1.236060 54.599837,-1.235639 54.599828,-1.235434 54.599864,-1.235202 54.599854,-1.235063 54.599822,-1.234897 54.599755,-1.234799 54.599734,-1.234647 54.599720,-1.234565 54.599690,-1.234484 54.599638,-1.234447 54.599608,-1.234405 54.599557,-1.234306 54.599471,-1.234267 54.599451,-1.234022 54.599388,-1.233858 54.599361,-1.233446 54.599375,-1.233442 54.599417,-1.233678 54.601451,-1.233948 54.603642,-1.236817 54.603639,-1.246748 54.603569,-1.250048 54.601986,-1.250198 54.601952,-1.248889 54.601045,-1.248357 54.600664,-1.247944 54.600349,-1.246529 54.600752,-1.245915 54.599941,-1.245832 54.599909,-1.245807 54.599911,-1.245686 54.599962,-1.245584 54.599994,-1.245013 54.600219,-1.244672 54.600364,-1.244425 54.600494,-1.244167 54.600571,-1.244048 54.600593,-1.244054 54.600605,-1.243709 54.600674,-1.243382 54.600820,-1.242351 54.601124,-1.242329 54.601101,-1.241955 54.600894,-1.241885 54.600845,-1.241862 54.600805,-1.241837 54.600792,-1.241697 54.600757,-1.241631 54.600749,-1.241580 54.600731,-1.241436 54.600726,-1.241406 54.600709,-1.241349 54.600658,-1.241331 54.600653,-1.241279 54.600595,-1.241261 54.600557,-1.241205 54.600525,-1.241081 54.600521,-1.240771 54.600548)))"  # noqa: E501
-    assert wkt.normalise(value, issues=issues) == expected
-
-
-def test_wkt_invalid_multipolygon_self_intersection_with_holes():
-    wkt = WktDataType()
-    issues = IssueLog()
-    # 56948 Wychavon District
-    # Self-intersection[-2.23750992455993 52.3608380047779]
-    value = "MULTIPOLYGON(((-2.219219 52.359138,-2.220211 52.359378,-2.220471 52.359422,-2.221685 52.359671,-2.223059 52.359918,-2.223321 52.359918,-2.224476 52.360181,-2.224934 52.360309,-2.225276 52.360421,-2.225417 52.360449,-2.226117 52.360489,-2.226391 52.360522,-2.230048 52.361069,-2.23038 52.361108,-2.230882 52.361148,-2.231879 52.361187,-2.232138 52.361205,-2.23311 52.361126,-2.233739 52.361018,-2.235523 52.360881,-2.23721 52.360857,-2.23751 52.360838,-2.237432 52.360688,-2.237512 52.360842,-2.237998 52.360783,-2.239279 52.360553,-2.240296 52.360428,-2.241611 52.360336,-2.241935 52.360368,-2.242738 52.360753,-2.242828 52.360787,-2.242966 52.360819,-2.244501 52.360918,-2.245688 52.361022,-2.246176 52.360063,-2.246586 52.35937,-2.246946 52.358795,-2.247101 52.358432,-2.246931 52.358449,-2.247499 52.357102,-2.247668 52.356678,-2.247808 52.35643,-2.247931 52.356144,-2.24825 52.355477,-2.248235 52.355473,-2.248248 52.355445,-2.247603 52.355271,-2.247506 52.355255,-2.247419 52.355261,-2.247287 52.355305,-2.245728 52.354927,-2.245735 52.354917,-2.245707 52.35491,-2.246763 52.353241,-2.24582 52.353017,-2.246116 52.352627,-2.247583 52.350831,-2.24769 52.350784,-2.249376 52.348833,-2.249439 52.348856,-2.249448 52.348846,-2.249802 52.348973,-2.249878 52.348847,-2.25007 52.348582,-2.250434 52.348079,-2.250952 52.3474,-2.251079 52.347203,-2.251191 52.347056,-2.251216 52.34701,-2.251659 52.345845,-2.251429 52.345849,-2.251097 52.34584,-2.250744 52.345802,-2.249882 52.345656,-2.24975 52.345653,-2.249651 52.345663,-2.249516 52.345665,-2.249537 52.34564,-2.249415 52.345652,-2.249481 52.345606,-2.249765 52.345473,-2.250106 52.345334,-2.251117 52.344948,-2.251371 52.344825,-2.251439 52.344783,-2.25175 52.34455,-2.252118 52.344229,-2.252137 52.344207,-2.252234 52.344028,-2.252422 52.34363,-2.252596 52.343282,-2.252628 52.343203,-2.252754 52.342968,-2.252872 52.342615,-2.253145 52.342118,-2.253206 52.342017,-2.253333 52.341845,-2.253542 52.341586,-2.253932 52.341125,-2.253981 52.341142,-2.254006 52.341115,-2.254735 52.341322,-2.255745 52.341556,-2.256257 52.341633,-2.256417 52.341647,-2.256684 52.341653,-2.256809 52.341679,-2.257019 52.341683,-2.257338 52.341658,-2.257543 52.341623,-2.257789 52.341559,-2.257946 52.341506,-2.260411 52.341036,-2.261065 52.340953,-2.261594 52.340866,-2.261983 52.340815,-2.262009 52.340743,-2.261232 52.340654,-2.261227 52.340648,-2.261243 52.340535,-2.261266 52.340461,-2.261301 52.340388,-2.261386 52.34026,-2.2617 52.339902,-2.26173 52.33983,-2.261737 52.339784,-2.261737 52.339728,-2.261725 52.33967,-2.261676 52.339575,-2.261611 52.339504,-2.261219 52.339217,-2.261089 52.339141,-2.260976 52.339101,-2.260919 52.33907,-2.260848 52.339005,-2.260709 52.338646,-2.260736 52.338406,-2.260872 52.338325,-2.261047 52.338248,-2.261181 52.338204,-2.261798 52.338058,-2.2631 52.338127,-2.263077 52.338016,-2.263068 52.337922,-2.263072 52.33783,-2.263285 52.336903,-2.263308 52.336841,-2.26337 52.336718,-2.263858 52.336039,-2.264079 52.335934,-2.26495 52.335971,-2.265184 52.335971,-2.265538 52.335952,-2.265868 52.335904,-2.265982 52.335899,-2.266097 52.335896,-2.266333 52.335924,-2.266826 52.33615,-2.267131 52.336275,-2.268065 52.336632,-2.268295 52.336425,-2.268478 52.336282,-2.26834 52.336212,-2.268366 52.336176,-2.268436 52.336108,-2.268478 52.336077,-2.268753 52.336065,-2.268876 52.336092,-2.268998 52.336126,-2.269216 52.336269,-2.26916 52.336135,-2.269161 52.336057,-2.269175 52.336,-2.269208 52.335931,-2.269235 52.335896,-2.269305 52.335841,-2.269229 52.335752,-2.269216 52.33571,-2.269213 52.335656,-2.26922 52.335626,-2.269269 52.33555,-2.269719 52.335152,-2.267975 52.334078,-2.267306 52.33416,-2.266837 52.333635,-2.267003 52.333569,-2.267208 52.333507,-2.267342 52.333512,-2.267833 52.333485,-2.267968 52.33374,-2.270334 52.333121,-2.271779 52.332618,-2.271123 52.331963,-2.270858 52.331634,-2.270783 52.331658,-2.270828 52.331784,-2.27082 52.331804,-2.270767 52.331862,-2.270514 52.332056,-2.270431 52.332094,-2.270346 52.332115,-2.27017 52.33208,-2.269903 52.331995,-2.269734 52.331928,-2.269379 52.331838,-2.268946 52.331646,-2.268627 52.331472,-2.268517 52.33138,-2.268476 52.331326,-2.268468 52.331296,-2.268376 52.331228,-2.267261 52.330812,-2.266464 52.330537,-2.266292 52.330461,-2.266156 52.33039,-2.265971 52.330276,-2.265553 52.330117,-2.265334 52.330051,-2.264944 52.329959,-2.264135 52.329802,-2.263841 52.329757,-2.263447 52.329637,-2.262763 52.329639,-2.262622 52.329625,-2.262261 52.329524,-2.261951 52.329446,-2.261659 52.329297,-2.261422 52.329153,-2.261161 52.328932,-2.260899 52.328771,-2.26025 52.328586,-2.259985 52.328528,-2.258908 52.328352,-2.258338 52.328205,-2.258204 52.328165,-2.257965 52.328072,-2.257672 52.327935,-2.257395 52.327784,-2.257207 52.327676,-2.256751 52.327382,-2.256239 52.327071,-2.255899 52.326878,-2.255135 52.3265,-2.254356 52.326155,-2.253899 52.325995,-2.251895 52.325501,-2.25161 52.325417,-2.251494 52.325375,-2.251266 52.325266,-2.251065 52.325164,-2.249734 52.324428,-2.249612 52.324343,-2.249415 52.324182,-2.249009 52.323791,-2.2485 52.323317,-2.24847 52.323311,-2.248384 52.323328,-2.248302 52.323356,-2.247721 52.323621,-2.246956 52.323937,-2.24681 52.32398,-2.246701 52.323996,-2.246417 52.324011,-2.246082 52.324003,-2.244877 52.323844,-2.244729 52.323833,-2.244618 52.323833,-2.24399 52.323888,-2.243637 52.323931,-2.242906 52.324036,-2.242683 52.324085,-2.242216 52.324138,-2.242029 52.324137,-2.241791 52.324096,-2.241159 52.324009,-2.240891 52.323975,-2.240744 52.32397,-2.240629 52.32399,-2.240582 52.324014,-2.240491 52.324088,-2.240427 52.324129,-2.240371 52.32415,-2.240193 52.324187,-2.240086 52.324194,-2.238681 52.324191,-2.238192 52.324197,-2.238066 52.324189,-2.23796 52.324173,-2.237192 52.324008,-2.237048 52.323984,-2.236532 52.32393,-2.236276 52.323914,-2.235472 52.323893,-2.235268 52.323905,-2.234935 52.323939,-2.233882 52.324114,-2.23339 52.324157,-2.233299 52.324152,-2.233133 52.324127,-2.232933 52.324082,-2.23268 52.32401,-2.232561 52.324002,-2.232418 52.324018,-2.231378 52.324023,-2.230386 52.32391,-2.229426 52.32386,-2.229168 52.323839,-2.228476 52.32378,-2.227666 52.323688,-2.226645 52.323599,-2.226303 52.323581,-2.22551 52.323555,-2.224674 52.323548,-2.222433 52.323576,-2.22177 52.323593,-2.219554 52.323676,-2.219186 52.323678,-2.218596 52.323663,-2.217935 52.323615,-2.217612 52.323579,-2.215772 52.32331,-2.215379 52.32326,-2.214933 52.323219,-2.214828 52.323091,-2.214618 52.322731,-2.214566 52.322655,-2.214195 52.322138,-2.213858 52.321599,-2.212909 52.320136,-2.212881 52.320139,-2.21287 52.320133,-2.212865 52.320115,-2.212891 52.320104,-2.212668 52.31976,-2.211414 52.318001,-2.211164 52.317585,-2.210969 52.317285,-2.210555 52.316763,-2.210414 52.316539,-2.21015 52.316157,-2.2097 52.315556,-2.209525 52.315315,-2.209492 52.31525,-2.209155 52.314907,-2.208656 52.314437,-2.208045 52.313948,-2.207645 52.313607,-2.207454 52.313477,-2.207151 52.313302,-2.206783 52.31311,-2.205236 52.312387,-2.204279 52.312002,-2.204041 52.311918,-2.204048 52.311909,-2.203804 52.311811,-2.20298 52.311505,-2.200164 52.310598,-2.199682 52.310487,-2.199274 52.310432,-2.198991 52.310363,-2.198039 52.310094,-2.197348 52.309915,-2.197041 52.309843,-2.195562 52.309458,-2.194698 52.309179,-2.194607 52.309171,-2.194566 52.309192,-2.194059 52.308965,-2.193892 52.308899,-2.193484 52.308773,-2.19273 52.308559,-2.192514 52.308476,-2.191466 52.308154,-2.190573 52.307851,-2.19003 52.307644,-2.189245 52.307362,-2.187706 52.306743,-2.187383 52.30663,-2.186503 52.306229,-2.186078 52.306025,-2.185728 52.305831,-2.185045 52.305505,-2.18475 52.305374,-2.184313 52.305215,-2.183867 52.305034,-2.183517 52.304835,-2.183496 52.304815,-2.183484 52.304788,-2.183208 52.304683,-2.182763 52.304487,-2.182488 52.30435,-2.18234 52.304298,-2.18224 52.304245,-2.182218 52.304224,-2.181076 52.303659,-2.18049 52.303383,-2.180005 52.303049,-2.179765 52.302866,-2.179494 52.302703,-2.17898 52.302352,-2.178363 52.301873,-2.177925 52.301577,-2.177816 52.301495,-2.177507 52.301243,-2.177305 52.301043,-2.177306 52.30103,-2.177284 52.30098,-2.177306 52.300952,-2.177323 52.300943,-2.177309 52.300924,-2.177062 52.300721,-2.176974 52.30067,-2.176712 52.300431,-2.176301 52.300013,-2.176175 52.299895,-2.175687 52.299377,-2.175661 52.299366,-2.175549 52.299267,-2.175285 52.299094,-2.173739 52.297686,-2.172665 52.296728,-2.172306 52.296391,-2.172212 52.296285,-2.172144 52.296189,-2.172054 52.296084,-2.171884 52.295898,-2.170878 52.294971,-2.17081 52.294911,-2.170793 52.294916,-2.169356 52.293535,-2.16871 52.292801,-2.168576 52.292688,-2.167765 52.291862,-2.167285 52.291402,-2.16691 52.29107,-2.166472 52.290739,-2.166227 52.290542,-2.165783 52.290218,-2.165184 52.289735,-2.164453 52.288998,-2.164423 52.288935,-2.164337 52.288937,-2.164137 52.288771,-2.163589 52.288274,-2.163177 52.287867,-2.162777 52.287424,-2.162721 52.287374,-2.162832 52.287344,-2.16284 52.287323,-2.162815 52.287287,-2.16256 52.287051,-2.16249 52.286994,-2.162381 52.286936,-2.162309 52.286888,-2.16177 52.286354,-2.161519 52.286127,-2.161303 52.285947,-2.161207 52.285869,-2.161132 52.28589,-2.161071 52.285863,-2.160828 52.285572,-2.16058 52.285254,-2.160223 52.284829,-2.160187 52.284799,-2.160082 52.28481,-2.160053 52.284772,-2.160069 52.284767,-2.160057 52.284745,-2.160098 52.284696,-2.159613 52.284357,-2.159406 52.284252,-2.159208 52.284046,-2.159238 52.28403,-2.158606 52.28354,-2.158363 52.283287,-2.15809 52.283076,-2.157663 52.282713,-2.157522 52.28258,-2.156648 52.281699,-2.156009 52.280972,-2.155863 52.280815,-2.155793 52.280761,-2.155724 52.280694,-2.155056 52.280027,-2.154887 52.279811,-2.154507 52.279234,-2.154377 52.279017,-2.15426 52.278803,-2.153952 52.278162,-2.153788 52.277774,-2.15357 52.277205,-2.153479 52.27674,-2.153344 52.275658,-2.153337 52.275562,-2.153342 52.275076,-2.153362 52.274834,-2.15336 52.274723,-2.153328 52.274689,-2.15334 52.274618,-2.153118 52.274561,-2.15304 52.274629,-2.152931 52.274662,-2.152261 52.274542,-2.15199 52.274509,-2.151777 52.274472,-2.151575 52.274421,-2.150223 52.274027,-2.150059 52.273941,-2.149812 52.27385,-2.149374 52.273737,-2.148982 52.273623,-2.148098 52.273385,-2.147926 52.273365,-2.147052 52.273224,-2.146156 52.273065,-2.145637 52.273027,-2.145325 52.273023,-2.145044 52.273049,-2.144518 52.273107,-2.144083 52.273197,-2.143846 52.273233,-2.142873 52.273311,-2.142723 52.27333,-2.140601 52.273627,-2.140123 52.273737,-2.139879 52.273321,-2.139849 52.273288,-2.139793 52.273089,-2.139801 52.273055,-2.139758 52.272902,-2.139874 52.272858,-2.139885 52.272842,-2.140452 52.272424,-2.140569 52.272322,-2.140819 52.272123,-2.141517 52.271585,-2.140963 52.271307,-2.140668 52.271093,-2.140684 52.270594,-2.140745 52.27059,-2.140742 52.270541,-2.140825 52.269835,-2.141008 52.269702,-2.141013 52.269601,-2.140619 52.269575,-2.140299 52.269534,-2.139733 52.269483,-2.139028 52.269498,-2.138386 52.269487,-2.138102 52.269489,-2.137727 52.269507,-2.137247 52.269544,-2.137088 52.269546,-2.136636 52.269532,-2.136455 52.269537,-2.135216 52.269605,-2.135134 52.26963,-2.135093 52.269673,-2.135086 52.269712,-2.135095 52.269774,-2.135035 52.269782,-2.135001 52.269717,-2.134906 52.269681,-2.13479 52.269683,-2.134423 52.269743,-2.134042 52.269837,-2.133267 52.269965,-2.132411 52.270121,-2.13191 52.270197,-2.131867 52.270103,-2.131861 52.270066,-2.131868 52.270052,-2.131906 52.270011,-2.131973 52.269966,-2.132107 52.269919,-2.132177 52.269857,-2.132205 52.269808,-2.13221 52.269778,-2.132204 52.269692,-2.13221 52.269664,-2.132194 52.269627,-2.132144 52.269583,-2.132144 52.26949,-2.132124 52.269451,-2.132132 52.269426,-2.132174 52.269395,-2.132297 52.26937,-2.132344 52.269353,-2.132385 52.269296,-2.132458 52.269243,-2.132819 52.269195,-2.132635 52.268913,-2.132521 52.268863,-2.131987 52.268576,-2.131654 52.268386,-2.132247 52.268106,-2.132677 52.267881,-2.133159 52.267691,-2.133525 52.267524,-2.133697 52.267436,-2.133681 52.267397,-2.133704 52.26737,-2.133801 52.267356,-2.13408 52.26721,-2.134263 52.267096,-2.134357 52.267027,-2.134536 52.266872,-2.134395 52.266809,-2.133882 52.266634,-2.133596 52.266567,-2.133274 52.266519,-2.132821 52.266488,-2.131841 52.26648,-2.131317 52.266456,-2.130986 52.266432,-2.13027 52.266281,-2.129602 52.266161,-2.129198 52.266117,-2.128694 52.266075,-2.128091 52.266053,-2.127364 52.266075,-2.126837 52.26612,-2.126522 52.266128,-2.126212 52.266119,-2.119952 52.265298,-2.115125 52.264694,-2.11148 52.264109,-2.111087 52.264058,-2.110021 52.2639,-2.109683 52.263867,-2.108934 52.263822,-2.107455 52.263701,-2.106395 52.263598,-2.105909 52.263562,-2.10332 52.263314,-2.101655 52.263172,-2.100815 52.26309,-2.09765 52.262809,-2.096722 52.262699,-2.093677 52.262226,-2.091837 52.262005,-2.090728 52.261882,-2.089397 52.261757,-2.086862 52.261558,-2.086106 52.261448,-2.083137 52.261193,-2.080816 52.260884,-2.07721 52.260601,-2.072849 52.260144,-2.071848 52.260094,-2.071465 52.260083,-2.070513 52.260085,-2.07004 52.260097,-2.069706 52.26011,-2.068625 52.260191,-2.067797 52.26029,-2.067344 52.260359,-2.066804 52.260453,-2.065932 52.260633,-2.061021 52.261689,-2.060301 52.261871,-2.059602 52.262071,-2.05719 52.262856,-2.05588 52.263301,-2.055105 52.26362,-2.054654 52.263835,-2.054334 52.264,-2.05265 52.264933,-2.054113 52.265903,-2.053168 52.266439,-2.053286 52.266516,-2.053487 52.266625,-2.053178 52.266876,-2.053024 52.266979,-2.052999 52.267033,-2.053277 52.267229,-2.053316 52.267264,-2.05305 52.267393,-2.053179 52.267468,-2.052755 52.267682,-2.053013 52.268006,-2.052842 52.268069,-2.052624 52.268163,-2.051959 52.268357,-2.051412 52.268538,-2.051342 52.268633,-2.051278 52.268704,-2.051134 52.268926,-2.050976 52.268815,-2.05082 52.268675,-2.050594 52.268437,-2.050524 52.268344,-2.050376 52.268077,-2.050344 52.26799,-2.050344 52.267943,-2.050273 52.267892,-2.050228 52.267839,-2.050292 52.267799,-2.050331 52.267732,-2.050466 52.267552,-2.0508 52.267331,-2.050845 52.267284,-2.050925 52.266893,-2.050903 52.266893,-2.051031 52.266529,-2.051008 52.266523,-2.051053 52.266391,-2.051079 52.266281,-2.051012 52.266034,-2.050335 52.265812,-2.049705 52.265702,-2.049912 52.2653,-2.049183 52.265186,-2.048167 52.264985,-2.048042 52.264947,-2.047824 52.264865,-2.047609 52.264775,-2.04704 52.264509,-2.046443 52.26428,-2.046032 52.264164,-2.04578 52.264114,-2.045098 52.263958,-2.044804 52.263875,-2.044171 52.263675,-2.043079 52.263288,-2.041948 52.262863,-2.040945 52.262503,-2.040575 52.262352,-2.040183 52.262177,-2.040023 52.262067,-2.039769 52.261928,-2.039461 52.261783,-2.038694 52.26147,-2.036469 52.260611,-2.035455 52.260186,-2.032303 52.258986,-2.030577 52.258311,-2.027489 52.257155,-2.026896 52.256947,-2.02659 52.256844,-2.026198 52.256724,-2.02562 52.256556,-2.025304 52.256473,-2.024781 52.256343,-2.022064 52.255722,-2.021195 52.255506,-2.02057 52.255342,-2.020223 52.255241,-2.01952 52.254997,-2.019093 52.254829,-2.018639 52.254628,-2.017926 52.254306,-2.017139 52.253933,-2.01591 52.253315,-2.014882 52.252767,-2.014409 52.253113,-2.014356 52.253125,-2.014104 52.253149,-2.013732 52.253224,-2.013712 52.253322,-2.013741 52.253391,-2.013744 52.253416,-2.013713 52.253546,-2.013652 52.25369,-2.013634 52.253755,-2.013501 52.25405,-2.013463 52.254048,-2.013403 52.254201,-2.013282 52.254588,-2.013248 52.254836,-2.013153 52.25516,-2.013133 52.255403,-2.013234 52.255501,-2.013272 52.255502,-2.013348 52.255523,-2.013531 52.255657,-2.013751 52.256016,-2.013817 52.25617,-2.013782 52.25648,-2.013789 52.256555,-2.013582 52.257011,-2.013461 52.25715,-2.013394 52.257278,-2.013359 52.25733,-2.013034 52.257735,-2.01334 52.257619,-2.013313 52.257656,-2.013385 52.257629,-2.013338 52.25768,-2.013151 52.257941,-2.013092 52.258006,-2.013095 52.258002,-2.013036 52.258,-2.01289 52.257954,-2.012852 52.257965,-2.012765 52.258104,-2.012738 52.258224,-2.012724 52.258248,-2.012591 52.258394,-2.0126 52.258479,-2.012552 52.258562,-2.012559 52.258645,-2.01241 52.258817,-2.012306 52.258914,-2.012231 52.258966,-2.012228 52.258983,-2.012253 52.259027,-2.012251 52.259043,-2.012234 52.259088,-2.012133 52.259274,-2.011988 52.259383,-2.011896 52.259463,-2.011451 52.259926,-2.011405 52.259997,-2.011409 52.260046,-2.011453 52.260145,-2.011436 52.260186,-2.011425 52.260246,-2.011444 52.260498,-2.01148 52.260613,-2.011548 52.26077,-2.01151 52.260964,-2.011516 52.261016,-2.011506 52.261126,-2.011521 52.261154,-2.011553 52.261272,-2.011532 52.261312,-2.011534 52.261374,-2.011552 52.261594,-2.011579 52.261796,-2.011581 52.261883,-2.011565 52.262038,-2.011526 52.262265,-2.011487 52.262351,-2.011412 52.262482,-2.01128 52.262589,-2.011262 52.262614,-2.011249 52.262674,-2.011248 52.262794,-2.011276 52.262881,-2.011426 52.262932,-2.011869 52.262983,-2.0121 52.263027,-2.012183 52.263056,-2.012175 52.263235,-2.012164 52.263313,-2.012164 52.263431,-2.012205 52.263667,-2.012152 52.263722,-2.012147 52.263746,-2.012163 52.263892,-2.012217 52.26404,-2.012227 52.264166,-2.01226 52.264297,-2.012264 52.264371,-2.012275 52.264403,-2.012591 52.264325,-2.013001 52.264177,-2.013054 52.264169,-2.013158 52.264179,-2.013187 52.264193,-2.01322 52.264214,-2.013286 52.2643,-2.013296 52.264519,-2.013334 52.264649,-2.013388 52.26472,-2.013477 52.264903,-2.013668 52.265134,-2.01372 52.265216,-2.013916 52.26542,-2.013923 52.265454,-2.013918 52.265566,-2.013993 52.265666,-2.01398 52.265676,-2.014 52.2657,-2.013966 52.265717,-2.013893 52.265737,-2.013819 52.265789,-2.013783 52.265824,-2.013754 52.265836,-2.013514 52.26591,-2.013339 52.265986,-2.013267 52.266036,-2.013246 52.266113,-2.013482 52.266265,-2.014269 52.266514,-2.014445 52.266553,-2.014498 52.266558,-2.014603 52.266586,-2.014808 52.266687,-2.014899 52.266748,-2.015227 52.267016,-2.015368 52.267141,-2.015425 52.267207,-2.015453 52.267277,-2.015451 52.26732,-2.015388 52.267461,-2.015382 52.26766,-2.015391 52.267671,-2.015378 52.267725,-2.015383 52.267764,-2.015418 52.267838,-2.015395 52.267873,-2.01541 52.268023,-2.015456 52.268177,-2.015443 52.268264,-2.015526 52.268443,-2.015562 52.268655,-2.015519 52.268767,-2.015499 52.269305,-2.015553 52.269538,-2.015538 52.269609,-2.015556 52.269713,-2.01555 52.269962,-2.015538 52.27001,-2.0153 52.270031,-2.014998 52.270037,-2.014797 52.270031,-2.014631 52.270046,-2.014363 52.270089,-2.014171 52.270104,-2.014101 52.270119,-2.013838 52.270183,-2.013746 52.270228,-2.013726 52.270357,-2.013701 52.270446,-2.013656 52.270462,-2.013412 52.27048,-2.013175 52.270473,-2.012945 52.270455,-2.01268 52.27042,-2.012137 52.270535,-2.012061 52.270569,-2.011535 52.270887,-2.011513 52.270908,-2.011293 52.2709,-2.011173 52.270907,-2.010471 52.270896,-2.009828 52.270896,-2.009388 52.270761,-2.008923 52.270602,-2.008777 52.270568,-2.008708 52.270523,-2.008582 52.270486,-2.008516 52.270485,-2.008015 52.270407,-2.007955 52.270408,-2.007854 52.270424,-2.007763 52.270422,-2.007177 52.270265,-2.007111 52.270269,-2.006789 52.270324,-2.006608 52.270378,-2.006566 52.270406,-2.006532 52.27045,-2.006443 52.270494,-2.006359 52.27055,-2.006148 52.270706,-2.006078 52.270737,-2.005848 52.270595,-2.005266 52.27109,-2.004388 52.271857,-2.004221 52.272012,-2.003228 52.273624,-2.003955 52.273842,-2.004791 52.274226,-2.00569 52.274575,-2.005878 52.27466,-2.005977 52.274711,-2.006112 52.274797,-2.006391 52.275022,-2.006553 52.275136,-2.007052 52.275373,-2.007026 52.275451,-2.006992 52.275478,-2.006867 52.275564,-2.006796 52.275602,-2.006716 52.275662,-2.006684 52.275663,-2.006665 52.275673,-2.006399 52.27567,-2.006143 52.275652,-2.00583 52.275642,-2.005412 52.275662,-2.004867 52.275877,-2.003809 52.276317,-2.003539 52.276423,-2.003186 52.276559,-2.0025 52.276792,-2.002582 52.276892,-2.002702 52.277066,-2.002837 52.277292,-2.002938 52.277488,-2.00311 52.277767,-2.003336 52.27801,-2.003635 52.278297,-2.003917 52.278591,-2.003909 52.278597,-2.003919 52.278608,-2.003466 52.278914,-2.002685 52.279416,-2.002505 52.279545,-2.002665 52.279522,-2.002815 52.279485,-2.003018 52.279448,-2.003227 52.279393,-2.003568 52.27932,-2.003733 52.279257,-2.003921 52.279229,-2.003987 52.279247,-2.004053 52.279305,-2.004126 52.279349,-2.004596 52.2795,-2.004716 52.27951,-2.004758 52.279494,-2.004825 52.279436,-2.005009 52.279218,-2.005023 52.279182,-2.005025 52.279114,-2.005007 52.279066,-2.005708 52.278795,-2.006203 52.278545,-2.006307 52.278485,-2.006395 52.27849,-2.006579 52.278526,-2.006643 52.278527,-2.006733 52.278553,-2.007261 52.278589,-2.007947 52.278518,-2.007983 52.278578,-2.007989 52.278618,-2.008017 52.278654,-2.008122 52.2786,-2.008213 52.278569,-2.00837 52.278477,-2.008906 52.278106,-2.009038 52.278075,-2.009076 52.278056,-2.009161 52.278046,-2.009232 52.278059,-2.009298 52.278084,-2.009384 52.278138,-2.009466 52.278171,-2.009536 52.278215,-2.00956 52.278221,-2.009602 52.278257,-2.009929 52.27846,-2.010008 52.278489,-2.010053 52.278517,-2.010259 52.278665,-2.010362 52.278729,-2.01093 52.27854,-2.011356 52.278422,-2.011701 52.278353,-2.011758 52.278355,-2.011893 52.278383,-2.012173 52.278412,-2.012446 52.278425,-2.01361 52.278473,-2.01438 52.278494,-2.014757 52.278519,-2.014887 52.278492,-2.015058 52.278502,-2.015488 52.278567,-2.015719 52.278582,-2.016674 52.27877,-2.016983 52.27882,-2.017044 52.278812,-2.017095 52.278816,-2.017218 52.278855,-2.017597 52.278908,-2.017876 52.278989,-2.018143 52.279011,-2.018843 52.279148,-2.018915 52.279164,-2.019167 52.279254,-2.019332 52.279303,-2.019723 52.279407,-2.020041 52.279527,-2.02023 52.279396,-2.020339 52.279157,-2.020442 52.279098,-2.020473 52.279097,-2.020564 52.279035,-2.020644 52.278855,-2.020841 52.278755,-2.020919 52.278664,-2.020937 52.278666,-2.020994 52.278719,-2.021047 52.278732,-2.021128 52.2787,-2.021161 52.27866,-2.021196 52.278644,-2.021301 52.278621,-2.021382 52.278565,-2.021611 52.278498,-2.021806 52.27849,-2.022201 52.278402,-2.022338 52.278355,-2.022439 52.278336,-2.02247 52.278316,-2.022581 52.278325,-2.022653 52.278307,-2.022738 52.278363,-2.022885 52.278411,-2.02291 52.278408,-2.023106 52.278325,-2.023263 52.278086,-2.023281 52.278113,-2.023313 52.278054,-2.023329 52.278047,-2.023387 52.278197,-2.023396 52.278284,-2.023461 52.278389,-2.023499 52.2784,-2.023599 52.27839,-2.023628 52.278408,-2.023625 52.278473,-2.023613 52.278497,-2.023587 52.278521,-2.023531 52.278549,-2.023433 52.278577,-2.023421 52.278642,-2.023618 52.278658,-2.02366 52.278656,-2.023742 52.278634,-2.023813 52.278631,-2.023826 52.278634,-2.023836 52.278655,-2.023825 52.278679,-2.023826 52.278699,-2.023918 52.27873,-2.023985 52.27872,-2.024014 52.278733,-2.024014 52.278762,-2.02397 52.278819,-2.023951 52.278832,-2.023897 52.278834,-2.023885 52.278846,-2.023917 52.278925,-2.023964 52.278989,-2.023957 52.27903,-2.023965 52.279057,-2.024024 52.279158,-2.024026 52.279195,-2.024008 52.279231,-2.024014 52.279279,-2.023961 52.279298,-2.023959 52.279319,-2.023992 52.279443,-2.024086 52.279561,-2.024209 52.279544,-2.024224 52.279567,-2.024263 52.279563,-2.024282 52.279579,-2.024296 52.279612,-2.024296 52.279671,-2.0244 52.279734,-2.024476 52.279822,-2.024447 52.27986,-2.024444 52.279878,-2.024457 52.279889,-2.024497 52.279896,-2.024521 52.279909,-2.024516 52.279989,-2.02456 52.27998,-2.024615 52.279987,-2.024705 52.280015,-2.024705 52.280035,-2.024664 52.280052,-2.02464 52.280078,-2.024678 52.280113,-2.02468 52.280126,-2.024725 52.280153,-2.024743 52.280172,-2.024799 52.280183,-2.024835 52.280202,-2.02484 52.280218,-2.024827 52.280298,-2.024832 52.280368,-2.024856 52.280376,-2.025076 52.280357,-2.025243 52.28033,-2.02595 52.280616,-2.026153 52.280647,-2.026219 52.280641,-2.02631 52.280712,-2.026505 52.280822,-2.026628 52.28096,-2.026869 52.280978,-2.027077 52.280958,-2.0273 52.280952,-2.02748 52.28096,-2.027622 52.280994,-2.027729 52.281031,-2.027948 52.281142,-2.028159 52.281188,-2.02822 52.281275,-2.028297 52.281338,-2.028338 52.281356,-2.028407 52.281356,-2.028442 52.281365,-2.02852 52.281444,-2.028578 52.281491,-2.028646 52.28153,-2.028874 52.281633,-2.029714 52.282096,-2.030078 52.282355,-2.030281 52.282461,-2.030501 52.282534,-2.030618 52.282552,-2.030967 52.282554,-2.031084 52.282538,-2.031442 52.282521,-2.031706 52.282521,-2.031773 52.282526,-2.03184 52.282549,-2.032201 52.282622,-2.032623 52.282664,-2.032788 52.282637,-2.033123 52.282637,-2.033135 52.282663,-2.033159 52.282664,-2.033206 52.282747,-2.033229 52.282837,-2.033331 52.282878,-2.033503 52.282831,-2.033858 52.282644,-2.033854 52.282634,-2.033868 52.282638,-2.033895 52.282624,-2.034407 52.282818,-2.034377 52.282837,-2.034396 52.28285,-2.034087 52.283053,-2.033945 52.283161,-2.034116 52.283238,-2.034345 52.283292,-2.034774 52.283439,-2.034812 52.283467,-2.035173 52.283597,-2.035322 52.283676,-2.035409 52.283757,-2.035388 52.283766,-2.035461 52.283847,-2.035463 52.283868,-2.035482 52.283898,-2.035579 52.284148,-2.035655 52.284296,-2.035982 52.284204,-2.035995 52.284212,-2.036023 52.284203,-2.036092 52.284259,-2.03723 52.28491,-2.037793 52.285224,-2.037847 52.28523,-2.037851 52.285241,-2.038491 52.285625,-2.039867 52.286008,-2.040465 52.286181,-2.040725 52.286264,-2.041175 52.286364,-2.041472 52.286455,-2.041636 52.28652,-2.042654 52.28707,-2.043018 52.287244,-2.043205 52.287312,-2.043435 52.287379,-2.044189 52.28758,-2.044492 52.287647,-2.045021 52.287711,-2.045283 52.287734,-2.045534 52.287771,-2.046079 52.287795,-2.046641 52.287796,-2.046863 52.287806,-2.047014 52.287823,-2.047079 52.287838,-2.047852 52.2877,-2.047892 52.28764,-2.048324 52.287533,-2.04886 52.287309,-2.048904 52.287313,-2.04895 52.287295,-2.048958 52.287316,-2.048966 52.287316,-2.049629 52.287195,-2.050866 52.287057,-2.051209 52.286957,-2.052545 52.286378,-2.052929 52.286177,-2.053228 52.28597,-2.053522 52.285785,-2.053954 52.285492,-2.054364 52.285265,-2.054472 52.285171,-2.054576 52.285133,-2.054587 52.285142,-2.055054 52.284908,-2.05549 52.284742,-2.055516 52.284719,-2.055515 52.284704,-2.05555 52.28469,-2.056196 52.284552,-2.056796 52.28444,-2.05751 52.284337,-2.057668 52.284298,-2.05782 52.284279,-2.058061 52.284275,-2.058129 52.284282,-2.058266 52.284314,-2.058447 52.284384,-2.058609 52.284383,-2.059424 52.284297,-2.059726 52.284239,-2.059722 52.284232,-2.060016 52.284168,-2.060405 52.284103,-2.060679 52.284171,-2.06085 52.284176,-2.060935 52.284155,-2.061002 52.284219,-2.061398 52.284438,-2.061905 52.284742,-2.062286 52.284864,-2.062883 52.284939,-2.063442 52.285044,-2.063825 52.285177,-2.064109 52.28526,-2.067122 52.286288,-2.06727 52.286325,-2.06777 52.286416,-2.068365 52.286557,-2.070442 52.287148,-2.070643 52.287224,-2.070829 52.287278,-2.071135 52.287398,-2.071137 52.287417,-2.071155 52.287426,-2.071139 52.287446,-2.071147 52.287524,-2.071307 52.287637,-2.071509 52.287721,-2.071695 52.287777,-2.071703 52.287769,-2.073369 52.288352,-2.074205 52.288621,-2.074508 52.288679,-2.075064 52.288727,-2.075535 52.288834,-2.075774 52.288904,-2.076256 52.289002,-2.076744 52.289064,-2.077055 52.289078,-2.078275 52.289072,-2.078724 52.289081,-2.079201 52.28905,-2.079657 52.288989,-2.080043 52.288919,-2.080861 52.288811,-2.081174 52.288777,-2.08146 52.28878,-2.081625 52.288793,-2.081854 52.288827,-2.082291 52.288854,-2.082956 52.288852,-2.083308 52.288815,-2.0843 52.288846,-2.085219 52.288895,-2.085807 52.288953,-2.086436 52.288991,-2.087592 52.289034,-2.088 52.289104,-2.088219 52.289151,-2.088924 52.28933,-2.089143 52.289416,-2.089221 52.289467,-2.08934 52.289473,-2.089449 52.289454,-2.089574 52.289408,-2.089747 52.28938,-2.089879 52.289347,-2.090188 52.289235,-2.090398 52.289134,-2.090905 52.288765,-2.091065 52.288682,-2.09121 52.288634,-2.091697 52.288546,-2.09169 52.288555,-2.091761 52.288544,-2.09167 52.288663,-2.091704 52.288665,-2.091599 52.288817,-2.091037 52.289604,-2.090062 52.29094,-2.08985 52.291267,-2.089621 52.291537,-2.089496 52.291702,-2.089538 52.29171,-2.089489 52.29175,-2.089442 52.291806,-2.089307 52.291815,-2.088881 52.292379,-2.087762 52.293936,-2.087498 52.294289,-2.086921 52.294974,-2.086501 52.29543,-2.086097 52.295818,-2.085506 52.296304,-2.085518 52.296388,-2.085349 52.296586,-2.084887 52.296939,-2.084896 52.296977,-2.084922 52.296957,-2.08505 52.297304,-2.085253 52.29775,-2.085339 52.297897,-2.085471 52.298248,-2.08561 52.298529,-2.085647 52.298691,-2.086319 52.298609,-2.086429 52.29893,-2.086468 52.298927,-2.086526 52.299067,-2.086638 52.299185,-2.087165 52.300618,-2.087247 52.300612,-2.087466 52.301197,-2.087445 52.301203,-2.087646 52.301548,-2.087754 52.301717,-2.087778 52.301713,-2.087846 52.301842,-2.087453 52.301933,-2.087575 52.302147,-2.08768 52.302359,-2.088059 52.302315,-2.088586 52.303393,-2.0886 52.303445,-2.088639 52.30349,-2.088184 52.303642,-2.088 52.303716,-2.087934 52.303726,-2.087872 52.303718,-2.087965 52.303825,-2.088306 52.303689,-2.088368 52.303689,-2.088613 52.303651,-2.089002 52.303535,-2.089168 52.303474,-2.089415 52.303412,-2.089627 52.303376,-2.08968 52.303381,-2.089731 52.303398,-2.089808 52.30346,-2.090061 52.303576,-2.090177 52.303638,-2.09024 52.303682,-2.090264 52.303716,-2.090305 52.303817,-2.090353 52.303896,-2.090424 52.304772,-2.090411 52.304924,-2.090368 52.304919,-2.090054 52.305007,-2.089957 52.305056,-2.089914 52.305133,-2.089795 52.305256,-2.089661 52.305412,-2.08957 52.30544,-2.089398 52.305612,-2.089353 52.305671,-2.089157 52.305807,-2.089142 52.305801,-2.08903 52.305938,-2.088884 52.306162,-2.088746 52.306409,-2.088712 52.306486,-2.088691 52.306593,-2.088621 52.306803,-2.088535 52.306999,-2.088504 52.307051,-2.088412 52.307428,-2.088075 52.308255,-2.088001 52.30846,-2.08795 52.308631,-2.087737 52.308893,-2.087422 52.3094,-2.087301 52.309685,-2.087312 52.309752,-2.087305 52.309787,-2.087246 52.309884,-2.087184 52.309943,-2.0871 52.310041,-2.08704 52.310142,-2.086968 52.31034,-2.086917 52.310438,-2.086682 52.310766,-2.086573 52.310978,-2.086508 52.311069,-2.086327 52.311251,-2.086261 52.311301,-2.085715 52.311598,-2.085498 52.311817,-2.08541 52.311928,-2.085356 52.312024,-2.085283 52.312273,-2.08525 52.312691,-2.085369 52.312772,-2.085511 52.312892,-2.085635 52.312967,-2.085903 52.313191,-2.085792 52.313234,-2.085779 52.313248,-2.085634 52.313303,-2.085483 52.313373,-2.085421 52.313383,-2.085171 52.313484,-2.085278 52.313556,-2.085376 52.313642,-2.085635 52.313981,-2.085685 52.314011,-2.085748 52.314078,-2.085976 52.314252,-2.086179 52.314389,-2.086336 52.314511,-2.087168 52.314448,-2.088603 52.314282,-2.090353 52.314114,-2.091637 52.313925,-2.09385 52.313566,-2.09511 52.313352,-2.095063 52.313396,-2.095114 52.31339,-2.095021 52.313455,-2.09498 52.313493,-2.094943 52.313623,-2.094876 52.313814,-2.094887 52.313851,-2.09486 52.313915,-2.094872 52.313916,-2.094711 52.314213,-2.0946 52.314373,-2.094547 52.314502,-2.094506 52.314562,-2.094396 52.314662,-2.094283 52.314789,-2.094096 52.315121,-2.094117 52.315247,-2.094077 52.315289,-2.09411 52.315364,-2.094138 52.3154,-2.094279 52.315329,-2.095189 52.314764,-2.095501 52.314583,-2.095649 52.314529,-2.096035 52.314452,-2.096696 52.314342,-2.09697 52.314221,-2.097872 52.313866,-2.098594 52.313625,-2.099012 52.313523,-2.099169 52.313505,-2.099232 52.313487,-2.099232 52.313467,-2.099424 52.313476,-2.100191 52.31342,-2.100623 52.313402,-2.101965 52.313385,-2.102478 52.313468,-2.102733 52.313542,-2.103398 52.313667,-2.105261 52.313981,-2.105966 52.314135,-2.106211 52.314167,-2.106413 52.314209,-2.107741 52.314432,-2.108032 52.314465,-2.10861 52.314513,-2.109551 52.314558,-2.110031 52.314569,-2.111369 52.314531,-2.112306 52.314414,-2.113812 52.313707,-2.114038 52.313612,-2.114489 52.313444,-2.114296 52.313163,-2.114314 52.313158,-2.114308 52.31315,-2.115067 52.313092,-2.115593 52.313089,-2.117168 52.313107,-2.11744 52.313136,-2.118038 52.31324,-2.119347 52.313587,-2.119356 52.313577,-2.119987 52.313738,-2.120187 52.313565,-2.120199 52.313568,-2.120218 52.313551,-2.120894 52.313779,-2.121083 52.313832,-2.121394 52.313953,-2.122043 52.314135,-2.122291 52.314189,-2.123042 52.314326,-2.12384 52.314433,-2.124091 52.314483,-2.123785 52.314935,-2.123592 52.315188,-2.123448 52.31532,-2.123084 52.315622,-2.122828 52.315855,-2.122568 52.316112,-2.122506 52.316178,-2.122463 52.316241,-2.122313 52.316491,-2.122203 52.316615,-2.122052 52.316746,-2.1219 52.316853,-2.121795 52.316916,-2.121592 52.317011,-2.121536 52.317025,-2.121396 52.317086,-2.121146 52.317227,-2.120886 52.317362,-2.120798 52.317433,-2.120798 52.317469,-2.120749 52.317496,-2.120662 52.317512,-2.120535 52.317578,-2.120101 52.317836,-2.119744 52.318092,-2.11965 52.318187,-2.119598 52.318265,-2.119505 52.318373,-2.118967 52.318735,-2.118809 52.31886,-2.118619 52.319043,-2.1185 52.31918,-2.11839 52.319273,-2.118112 52.319549,-2.118039 52.319611,-2.117644 52.320018,-2.117261 52.320476,-2.117036 52.320778,-2.116937 52.320929,-2.116827 52.321063,-2.116742 52.321188,-2.116534 52.321428,-2.116464 52.32152,-2.116384 52.321592,-2.116 52.321856,-2.115902 52.321942,-2.115765 52.322085,-2.116078 52.322157,-2.116434 52.322224,-2.116631 52.322271,-2.117004 52.322374,-2.117281 52.322427,-2.117917 52.322589,-2.118185 52.322664,-2.118394 52.322734,-2.117899 52.323001,-2.117947 52.323041,-2.118041 52.323157,-2.118106 52.323314,-2.118155 52.323591,-2.118218 52.323777,-2.11824 52.323921,-2.118233 52.324028,-2.118283 52.324157,-2.118291 52.324285,-2.118346 52.3244,-2.118501 52.32444,-2.118493 52.324518,-2.118501 52.324587,-2.118564 52.324878,-2.118597 52.325156,-2.118346 52.325172,-2.117312 52.325162,-2.117316 52.325252,-2.117441 52.325744,-2.117491 52.326015,-2.117604 52.326394,-2.117641 52.326557,-2.11758 52.326613,-2.11748 52.326651,-2.117369 52.326711,-2.117307 52.32677,-2.117239 52.32686,-2.117114 52.327112,-2.117055 52.327253,-2.11662 52.3271,-2.116181 52.327087,-2.116224 52.3273,-2.116294 52.327498,-2.116278 52.327558,-2.116291 52.327653,-2.116331 52.327785,-2.116384 52.327897,-2.116542 52.328152,-2.116882 52.32857,-2.117014 52.328778,-2.117264 52.329278,-2.117316 52.329461,-2.117354 52.329645,-2.117384 52.329733,-2.117488 52.329972,-2.117758 52.330416,-2.11811 52.330938,-2.118206 52.331134,-2.118226 52.331141,-2.118225 52.331165,-2.118358 52.331372,-2.118363 52.331398,-2.118379 52.331397,-2.118375 52.331488,-2.118385 52.331538,-2.118472 52.331748,-2.118489 52.331769,-2.118546 52.331778,-2.118562 52.332037,-2.118548 52.332168,-2.11853 52.332208,-2.118422 52.332358,-2.11833 52.33252,-2.118086 52.332969,-2.11798 52.333191,-2.117965 52.333239,-2.117942 52.333486,-2.117849 52.333487,-2.117661 52.334663,-2.117608 52.335096,-2.117569 52.335274,-2.117546 52.335273,-2.11754 52.335316,-2.117373 52.335171,-2.11724 52.335168,-2.117218 52.335205,-2.117228 52.335228,-2.117228 52.335275,-2.117221 52.335299,-2.117201 52.335321,-2.117154 52.335325,-2.117098 52.335318,-2.117017 52.335328,-2.116976 52.335343,-2.11694 52.335381,-2.116868 52.335416,-2.11683 52.335413,-2.116804 52.335399,-2.116698 52.335407,-2.116693 52.335414,-2.116703 52.335473,-2.116696 52.335492,-2.116664 52.33551,-2.116478 52.335524,-2.116427 52.335519,-2.116249 52.335525,-2.11623 52.335505,-2.116232 52.335489,-2.11611 52.335417,-2.116068 52.335412,-2.116055 52.335402,-2.115825 52.335409,-2.115742 52.335429,-2.115672 52.335421,-2.115665 52.33528,-2.11546 52.335294,-2.115409 52.335285,-2.115186 52.335294,-2.115175 52.33529,-2.115169 52.335271,-2.115181 52.335243,-2.115239 52.335213,-2.115256 52.335194,-2.115236 52.335114,-2.115098 52.335147,-2.114971 52.335258,-2.114736 52.335307,-2.114681 52.335345,-2.11464 52.335391,-2.114588 52.335417,-2.114543 52.335468,-2.114559 52.335485,-2.114614 52.335516,-2.114636 52.335543,-2.114639 52.335576,-2.114615 52.335609,-2.114564 52.33562,-2.114479 52.335593,-2.114428 52.335593,-2.114381 52.335646,-2.114379 52.335662,-2.11436 52.335692,-2.114278 52.335773,-2.114223 52.33585,-2.11419 52.335861,-2.11412 52.335866,-2.114055 52.335902,-2.114044 52.335938,-2.113947 52.335956,-2.113757 52.336072,-2.113726 52.336142,-2.113689 52.336195,-2.113704 52.336237,-2.113698 52.336268,-2.113685 52.33628,-2.113647 52.336287,-2.113596 52.33628,-2.113521 52.336304,-2.113583 52.336368,-2.113633 52.336388,-2.113644 52.336412,-2.113481 52.33659,-2.113475 52.336669,-2.113491 52.336691,-2.113487 52.336715,-2.113456 52.336739,-2.11339 52.336745,-2.113322 52.336783,-2.113356 52.336805,-2.113431 52.336821,-2.113468 52.336837,-2.11348 52.336849,-2.11348 52.336868,-2.113464 52.336896,-2.113414 52.336903,-2.11334 52.336901,-2.113275 52.336939,-2.113295 52.336954,-2.113388 52.33697,-2.113404 52.336989,-2.113401 52.337003,-2.113347 52.337045,-2.11329 52.337147,-2.113485 52.337307,-2.113653 52.337381,-2.113698 52.337428,-2.113746 52.337493,-2.113884 52.337566,-2.113909 52.337594,-2.113941 52.337653,-2.113966 52.337673,-2.114067 52.337718,-2.114145 52.337727,-2.114178 52.337746,-2.11419 52.337832,-2.114212 52.337898,-2.114217 52.337987,-2.114233 52.338019,-2.114222 52.338057,-2.114226 52.338119,-2.114251 52.338195,-2.114291 52.338236,-2.114311 52.338326,-2.114311 52.338398,-2.114303 52.338432,-2.114275 52.338475,-2.114226 52.338511,-2.114293 52.338524,-2.114817 52.338569,-2.115906 52.338688,-2.116485 52.338732,-2.116374 52.338905,-2.116333 52.339069,-2.116337 52.339161,-2.116373 52.339164,-2.116371 52.339209,-2.116426 52.339287,-2.11638 52.339683,-2.116341 52.339791,-2.116303 52.339857,-2.116229 52.339953,-2.116096 52.340166,-2.11678 52.340098,-2.116871 52.340364,-2.116961 52.340546,-2.117126 52.340935,-2.117432 52.341606,-2.117443 52.341691,-2.117415 52.341729,-2.117385 52.341744,-2.117268 52.341752,-2.117037 52.341785,-2.117002 52.341924,-2.117105 52.341972,-2.117333 52.342054,-2.118257 52.342166,-2.118881 52.342171,-2.118937 52.342077,-2.118984 52.342032,-2.119057 52.341923,-2.119173 52.341702,-2.119199 52.341644,-2.119268 52.341429,-2.119387 52.34142,-2.119643 52.34134,-2.119882 52.341209,-2.120199 52.341051,-2.120634 52.340809,-2.121048 52.340624,-2.120996 52.340551,-2.120897 52.34045,-2.121397 52.340281,-2.121708 52.340214,-2.121866 52.340194,-2.12312 52.340106,-2.123918 52.34006,-2.123903 52.340207,-2.124721 52.340227,-2.124719 52.341006,-2.124758 52.341979,-2.123875 52.342144,-2.122818 52.342358,-2.122811 52.342383,-2.122819 52.342411,-2.122897 52.342504,-2.123337 52.342965,-2.123454 52.343075,-2.12361 52.343154,-2.123718 52.34324,-2.123991 52.343389,-2.124168 52.343458,-2.124411 52.34352,-2.124646 52.343606,-2.124852 52.343706,-2.125255 52.343882,-2.125366 52.343948,-2.125388 52.343981,-2.125391 52.344009,-2.125449 52.344109,-2.125655 52.344338,-2.125871 52.34456,-2.126069 52.344858,-2.126167 52.344971,-2.126248 52.345095,-2.126334 52.345199,-2.126366 52.345279,-2.126384 52.345361,-2.126381 52.345506,-2.126389 52.345539,-2.126779 52.345507,-2.127042 52.345475,-2.127229 52.345471,-2.127828 52.345479,-2.128443 52.345499,-2.128745 52.345543,-2.128808 52.345568,-2.129027 52.345606,-2.129128 52.345656,-2.129192 52.345675,-2.129368 52.345771,-2.130014 52.346279,-2.130254 52.346498,-2.130533 52.34668,-2.130844 52.346852,-2.131172 52.34695,-2.131444 52.347043,-2.13144 52.347055,-2.131477 52.347065,-2.131341 52.347236,-2.131021 52.347695,-2.130887 52.347865,-2.13086 52.347931,-2.130776 52.348005,-2.130291 52.348599,-2.129911 52.349107,-2.129556 52.34962,-2.129539 52.34968,-2.129483 52.349796,-2.129311 52.350279,-2.129222 52.350434,-2.129199 52.350498,-2.129157 52.350548,-2.129092 52.350662,-2.129052 52.350669,-2.12902 52.350716,-2.128995 52.350709,-2.128874 52.350863,-2.128699 52.351036,-2.128599 52.351115,-2.128469 52.351209,-2.128008 52.351516,-2.127638 52.351729,-2.127536 52.351798,-2.127331 52.35196,-2.127747 52.35223,-2.128179 52.352209,-2.128336 52.352217,-2.12859 52.352213,-2.128747 52.352224,-2.129133 52.352198,-2.129363 52.352099,-2.129477 52.352079,-2.12979 52.351987,-2.130048 52.351787,-2.130735 52.351314,-2.131491 52.350616,-2.131605 52.350491,-2.131622 52.350496,-2.131775 52.350338,-2.131806 52.350284,-2.132012 52.350102,-2.132009 52.350091,-2.132111 52.349876,-2.132377 52.349221,-2.132424 52.349067,-2.132484 52.348928,-2.132765 52.348159,-2.133023 52.347998,-2.133131 52.348034,-2.133156 52.348025,-2.133893 52.348442,-2.13469 52.348882,-2.135149 52.349076,-2.135641 52.349317,-2.135723 52.349369,-2.136057 52.349531,-2.136129 52.349545,-2.136206 52.349579,-2.136264 52.349595,-2.136565 52.349735,-2.137112 52.349942,-2.137173 52.349921,-2.137261 52.349935,-2.137514 52.349922,-2.137672 52.349843,-2.138014 52.349513,-2.138083 52.349432,-2.138199 52.349255,-2.138278 52.349111,-2.138421 52.348883,-2.138595 52.3487,-2.138891 52.348503,-2.139245 52.348351,-2.139602 52.348288,-2.139885 52.348127,-2.140037 52.3481,-2.140074 52.34808,-2.140119 52.348041,-2.140146 52.348031,-2.140204 52.348031,-2.140268 52.348016,-2.140483 52.34802,-2.141926 52.347937,-2.142056 52.347918,-2.142103 52.347669,-2.142113 52.347657,-2.142114 52.347581,-2.142129 52.347542,-2.142187 52.347471,-2.142315 52.347366,-2.142348 52.347324,-2.142421 52.347199,-2.142626 52.346985,-2.142704 52.346914,-2.142816 52.346874,-2.143331 52.346642,-2.143822 52.346436,-2.144303 52.346293,-2.144772 52.346065,-2.145786 52.345503,-2.146264 52.345228,-2.146698 52.345411,-2.146962 52.345505,-2.147218 52.345624,-2.148219 52.346181,-2.148509 52.34638,-2.148608 52.346416,-2.148864 52.346294,-2.148891 52.34627,-2.148974 52.346107,-2.148978 52.346081,-2.149007 52.346025,-2.14903 52.346006,-2.149039 52.345978,-2.149108 52.34594,-2.14916 52.345896,-2.149181 52.345844,-2.149203 52.345818,-2.149257 52.345777,-2.149298 52.345645,-2.149317 52.345612,-2.149374 52.345575,-2.149548 52.3454,-2.149759 52.345253,-2.149709 52.3452,-2.149683 52.345156,-2.149675 52.345096,-2.149764 52.345011,-2.14976 52.344998,-2.149779 52.344975,-2.149772 52.344967,-2.149701 52.344967,-2.149799 52.34473,-2.149823 52.344689,-2.149891 52.344633,-2.149984 52.344584,-2.150003 52.34454,-2.150047 52.344485,-2.150144 52.344423,-2.150375 52.344195,-2.150561 52.344025,-2.150953 52.343686,-2.151118 52.343234,-2.15117 52.342818,-2.151196 52.342705,-2.15123 52.34261,-2.151266 52.342542,-2.151337 52.342437,-2.151588 52.342507,-2.151707 52.342175,-2.151792 52.341803,-2.151815 52.341739,-2.151839 52.341705,-2.151912 52.341632,-2.151961 52.341521,-2.151947 52.341479,-2.151954 52.341336,-2.151931 52.341175,-2.151925 52.341055,-2.151935 52.34101,-2.15201 52.34082,-2.151978 52.340706,-2.151973 52.340636,-2.15199 52.340528,-2.151986 52.340446,-2.152005 52.340384,-2.152052 52.340328,-2.152141 52.340091,-2.152411 52.339611,-2.152532 52.33923,-2.152599 52.338781,-2.152571 52.338687,-2.152361 52.338695,-2.152306 52.338704,-2.152303 52.338697,-2.152141 52.338125,-2.152111 52.337983,-2.152015 52.337841,-2.152023 52.337755,-2.151975 52.337663,-2.151921 52.337607,-2.151889 52.337587,-2.151896 52.337568,-2.151882 52.337481,-2.151865 52.337455,-2.15183 52.337423,-2.151824 52.337399,-2.151841 52.337305,-2.151995 52.337192,-2.151927 52.337036,-2.151925 52.33699,-2.151919 52.336985,-2.152022 52.336898,-2.152214 52.336713,-2.152252 52.336693,-2.152416 52.336642,-2.152624 52.336476,-2.152681 52.336407,-2.152807 52.336293,-2.15285 52.336239,-2.152925 52.336192,-2.153408 52.335725,-2.153758 52.335705,-2.153848 52.335684,-2.153882 52.335665,-2.154056 52.335496,-2.154262 52.335348,-2.154398 52.335207,-2.154458 52.335126,-2.154538 52.335043,-2.154581 52.334941,-2.154721 52.334768,-2.154731 52.334667,-2.154844 52.334455,-2.15482 52.334451,-2.154831 52.334421,-2.15482 52.334377,-2.154787 52.334351,-2.154813 52.334327,-2.154816 52.334309,-2.154831 52.33431,-2.154856 52.334279,-2.154961 52.334211,-2.155008 52.334142,-2.155018 52.334117,-2.155022 52.334047,-2.154978 52.333982,-2.154989 52.333945,-2.15501 52.333923,-2.155272 52.333816,-2.155285 52.333801,-2.155303 52.333735,-2.155366 52.333674,-2.155509 52.333581,-2.155527 52.333581,-2.155645 52.333636,-2.155684 52.333644,-2.155755 52.33364,-2.15583 52.333625,-2.155951 52.333567,-2.156103 52.333468,-2.156158 52.33344,-2.1562 52.33339,-2.156225 52.333349,-2.156251 52.333274,-2.15625 52.333219,-2.156205 52.333147,-2.156215 52.333124,-2.156364 52.332974,-2.156488 52.332837,-2.156512 52.332801,-2.156415 52.332365,-2.156256 52.331969,-2.156245 52.331759,-2.156167 52.331527,-2.156074 52.331323,-2.156055 52.331098,-2.156072 52.331021,-2.156137 52.330972,-2.156224 52.331001,-2.15652 52.331045,-2.156667 52.331058,-2.156922 52.331058,-2.15715 52.331034,-2.157615 52.330952,-2.158323 52.330782,-2.159536 52.330378,-2.159914 52.330213,-2.160154 52.329931,-2.160362 52.329721,-2.160482 52.329617,-2.160716 52.329441,-2.161531 52.32878,-2.161775 52.32857,-2.161811 52.3285,-2.161991 52.328437,-2.162078 52.328382,-2.162136 52.32833,-2.16223 52.328273,-2.162497 52.328146,-2.162617 52.328049,-2.162683 52.327981,-2.162832 52.327858,-2.163012 52.327683,-2.163238 52.327495,-2.163504 52.327221,-2.163642 52.327119,-2.163685 52.327063,-2.163736 52.326959,-2.163755 52.326934,-2.163814 52.326878,-2.163904 52.326818,-2.163969 52.326719,-2.164095 52.32658,-2.164155 52.326499,-2.164205 52.326405,-2.164244 52.326308,-2.164321 52.326195,-2.164351 52.326137,-2.164361 52.326091,-2.164382 52.326064,-2.164382 52.326039,-2.16434 52.326017,-2.164461 52.32594,-2.164749 52.325733,-2.165003 52.325616,-2.1652 52.325552,-2.165764 52.325421,-2.165973 52.325348,-2.165961 52.325286,-2.166443 52.325076,-2.166552 52.325007,-2.166641 52.324964,-2.16693 52.324774,-2.16705 52.324727,-2.167042 52.324712,-2.167168 52.324656,-2.167251 52.32463,-2.167556 52.324508,-2.167583 52.324483,-2.167617 52.324433,-2.167637 52.324384,-2.167635 52.324333,-2.167643 52.324309,-2.167688 52.324207,-2.167747 52.324111,-2.167768 52.323804,-2.167766 52.323644,-2.167805 52.323279,-2.167736 52.323146,-2.167752 52.322988,-2.167681 52.322773,-2.167652 52.322712,-2.167619 52.322562,-2.167602 52.322261,-2.167597 52.321704,-2.167563 52.321542,-2.167577 52.321386,-2.167565 52.321306,-2.167577 52.321226,-2.167554 52.321062,-2.167527 52.320962,-2.167557 52.320904,-2.16755 52.320719,-2.167564 52.320559,-2.167516 52.320391,-2.167507 52.320309,-2.167524 52.320227,-2.167551 52.320153,-2.167613 52.320029,-2.167707 52.319887,-2.167871 52.319691,-2.167837 52.319551,-2.167869 52.319431,-2.16787 52.319341,-2.167855 52.319293,-2.167868 52.319207,-2.167815 52.319076,-2.167829 52.318983,-2.16782 52.318983,-2.167827 52.318943,-2.167818 52.318923,-2.167807 52.318911,-2.167724 52.318873,-2.167593 52.31883,-2.167565 52.318812,-2.167534 52.31877,-2.167461 52.318567,-2.16719 52.318458,-2.167019 52.318429,-2.166805 52.318374,-2.166845 52.318353,-2.166812 52.318346,-2.166816 52.318337,-2.166873 52.31828,-2.167197 52.318118,-2.167337 52.318059,-2.167626 52.317981,-2.167828 52.317841,-2.167966 52.317805,-2.168157 52.31772,-2.168325 52.317613,-2.168364 52.317579,-2.168428 52.317485,-2.168522 52.317403,-2.168578 52.317366,-2.168703 52.317298,-2.168802 52.317259,-2.169007 52.317203,-2.16908 52.317187,-2.169202 52.317176,-2.169368 52.317183,-2.169461 52.31717,-2.169515 52.317156,-2.169651 52.317066,-2.169737 52.317055,-2.170013 52.317096,-2.170343 52.317102,-2.170484 52.317114,-2.170837 52.317174,-2.170985 52.31719,-2.171361 52.31727,-2.171475 52.317279,-2.171796 52.31728,-2.172001 52.317267,-2.172172 52.317218,-2.172637 52.317211,-2.172726 52.317221,-2.172919 52.317223,-2.17317 52.317205,-2.173231 52.317191,-2.173338 52.31709,-2.173571 52.316962,-2.173787 52.316803,-2.173968 52.316718,-2.174129 52.316665,-2.174429 52.316603,-2.174585 52.3166,-2.174595 52.316639,-2.174623 52.31664,-2.174671 52.31682,-2.17469 52.316962,-2.174903 52.317702,-2.174976 52.317861,-2.175098 52.318394,-2.175111 52.318512,-2.175105 52.318723,-2.175137 52.318864,-2.175171 52.318955,-2.175217 52.319173,-2.175312 52.319823,-2.175281 52.320017,-2.175092 52.320602,-2.174953 52.320846,-2.175157 52.320862,-2.175128 52.320925,-2.175161 52.320934,-2.175087 52.321118,-2.175078 52.321188,-2.175088 52.321285,-2.175086 52.321344,-2.175104 52.321455,-2.175203 52.321838,-2.175208 52.322077,-2.175234 52.322194,-2.175235 52.322248,-2.175192 52.322461,-2.175148 52.322798,-2.175018 52.323163,-2.174992 52.323374,-2.174922 52.323605,-2.174831 52.323865,-2.174762 52.324124,-2.174732 52.324263,-2.174716 52.324379,-2.174706 52.324609,-2.174627 52.324921,-2.174568 52.325072,-2.174475 52.325273,-2.174345 52.32565,-2.17424 52.325909,-2.174163 52.326047,-2.174062 52.326181,-2.174041 52.326226,-2.174034 52.326264,-2.173781 52.326794,-2.173714 52.326955,-2.173621 52.327099,-2.17344 52.327322,-2.173443 52.327399,-2.173305 52.327594,-2.173265 52.327665,-2.173226 52.327711,-2.173188 52.327735,-2.173136 52.327797,-2.173132 52.327808,-2.17316 52.327839,-2.17306 52.327966,-2.172849 52.328103,-2.172617 52.32822,-2.172411 52.328262,-2.172314 52.328291,-2.172239 52.328334,-2.17219 52.328381,-2.172082 52.328536,-2.172051 52.328651,-2.172027 52.328704,-2.171989 52.328767,-2.171897 52.328879,-2.171729 52.329352,-2.171692 52.32948,-2.17162 52.329608,-2.171626 52.32961,-2.171556 52.329714,-2.171551 52.329734,-2.171225 52.330148,-2.171069 52.330395,-2.170947 52.330645,-2.170744 52.331426,-2.170772 52.331516,-2.170925 52.331642,-2.17109 52.331704,-2.171152 52.331716,-2.171202 52.331718,-2.171623 52.331686,-2.172261 52.331684,-2.172524 52.331692,-2.172665 52.331706,-2.172735 52.331707,-2.172817 52.331696,-2.172994 52.331713,-2.17322 52.331718,-2.173343 52.331713,-2.17342 52.331723,-2.17366 52.331728,-2.173927 52.331666,-2.17404 52.331648,-2.174143 52.331452,-2.174194 52.331397,-2.174286 52.331356,-2.174374 52.331343,-2.175252 52.331355,-2.175561 52.331392,-2.175703 52.331399,-2.175772 52.331424,-2.175827 52.331471,-2.176049 52.331552,-2.176318 52.331621,-2.176565 52.33166,-2.176894 52.331696,-2.177422 52.3317,-2.177705 52.331757,-2.177882 52.331756,-2.178003 52.331702,-2.178084 52.331638,-2.178135 52.331583,-2.178189 52.331555,-2.178264 52.331495,-2.178471 52.331432,-2.17858 52.331417,-2.178718 52.331377,-2.178949 52.331337,-2.179223 52.331255,-2.179737 52.331181,-2.180123 52.33114,-2.180877 52.3311,-2.181001 52.331085,-2.181238 52.331034,-2.181453 52.330966,-2.181606 52.330892,-2.182079 52.330633,-2.182259 52.330575,-2.182558 52.330432,-2.182666 52.330371,-2.182889 52.330308,-2.182955 52.330296,-2.183352 52.330261,-2.183495 52.330224,-2.183557 52.330171,-2.183584 52.330159,-2.184212 52.330135,-2.184803 52.3301,-2.184958 52.330076,-2.185085 52.330038,-2.185572 52.329844,-2.185688 52.329784,-2.185824 52.329743,-2.18606 52.329691,-2.186245 52.32961,-2.186349 52.329575,-2.186471 52.329541,-2.186729 52.329489,-2.187065 52.329331,-2.187199 52.329225,-2.18732 52.32918,-2.187527 52.329169,-2.187637 52.329201,-2.187696 52.329209,-2.188117 52.329218,-2.188283 52.329183,-2.188438 52.329176,-2.189051 52.329076,-2.189045 52.329056,-2.189104 52.329048,-2.189338 52.329087,-2.189494 52.329083,-2.189647 52.329019,-2.189885 52.328899,-2.190011 52.328871,-2.190265 52.328875,-2.190386 52.328899,-2.190497 52.328932,-2.190507 52.329013,-2.190498 52.329242,-2.190519 52.329528,-2.190581 52.329874,-2.190581 52.329917,-2.190553 52.330558,-2.190472 52.330868,-2.190469 52.330901,-2.190495 52.331033,-2.190495 52.331106,-2.190352 52.33154,-2.190358 52.331598,-2.190444 52.331875,-2.19048 52.33195,-2.190545 52.332041,-2.190718 52.332205,-2.190914 52.332367,-2.191054 52.33253,-2.191102 52.332575,-2.191222 52.332684,-2.191465 52.332881,-2.191563 52.333012,-2.191623 52.333049,-2.191964 52.33316,-2.192041 52.333194,-2.192121 52.333247,-2.192186 52.333301,-2.192457 52.33345,-2.192489 52.333476,-2.192519 52.33351,-2.192538 52.33355,-2.192577 52.333665,-2.19265 52.333936,-2.192656 52.333982,-2.192629 52.33414,-2.192633 52.334196,-2.192594 52.334518,-2.192606 52.334542,-2.192672 52.334617,-2.192913 52.33478,-2.19298 52.3348,-2.193103 52.334792,-2.193148 52.334798,-2.193198 52.334821,-2.193209 52.334838,-2.193197 52.334894,-2.193146 52.334988,-2.193161 52.335029,-2.193232 52.335111,-2.193332 52.335182,-2.193382 52.335231,-2.19348 52.335262,-2.193494 52.335283,-2.193492 52.335308,-2.193485 52.335314,-2.193375 52.33535,-2.193394 52.335391,-2.193481 52.335475,-2.193496 52.335502,-2.19349 52.335517,-2.193438 52.335561,-2.193321 52.335581,-2.193233 52.335622,-2.193227 52.335645,-2.193245 52.335709,-2.19324 52.33577,-2.193258 52.335842,-2.193197 52.336003,-2.193185 52.336087,-2.193176 52.336102,-2.193133 52.336109,-2.193044 52.336084,-2.192998 52.336066,-2.192849 52.335968,-2.192791 52.335988,-2.192762 52.336014,-2.192753 52.336049,-2.192763 52.336073,-2.192833 52.336148,-2.192827 52.336163,-2.192758 52.336242,-2.19259 52.336367,-2.19252 52.336457,-2.192513 52.336479,-2.192519 52.336494,-2.19261 52.336561,-2.192661 52.33661,-2.192739 52.336623,-2.19281 52.336607,-2.192871 52.33661,-2.192877 52.336625,-2.192866 52.336642,-2.19274 52.336681,-2.192709 52.336711,-2.192727 52.33676,-2.192809 52.336855,-2.192832 52.336891,-2.192832 52.33691,-2.192821 52.336918,-2.19278 52.336921,-2.192772 52.336931,-2.192492 52.336888,-2.192412 52.33695,-2.192409 52.336973,-2.192212 52.336915,-2.192173 52.336927,-2.192116 52.337023,-2.1921 52.337094,-2.192034 52.337152,-2.192005 52.337191,-2.192008 52.337208,-2.192023 52.337221,-2.192082 52.337246,-2.192107 52.33727,-2.192104 52.337281,-2.192031 52.337349,-2.192023 52.337366,-2.192019 52.337412,-2.19204 52.337524,-2.192052 52.337543,-2.192075 52.337556,-2.192341 52.337458,-2.192405 52.337464,-2.192448 52.337476,-2.192485 52.337496,-2.192514 52.337529,-2.192535 52.337541,-2.192539 52.337555,-2.192529 52.337621,-2.192518 52.337639,-2.192477 52.337671,-2.192399 52.337708,-2.192383 52.337726,-2.192398 52.337765,-2.19246 52.337833,-2.192472 52.337843,-2.192574 52.337809,-2.192678 52.337817,-2.192722 52.33781,-2.192874 52.337903,-2.192996 52.337918,-2.193027 52.33793,-2.19304 52.337946,-2.193093 52.338026,-2.193088 52.338068,-2.193058 52.338091,-2.192932 52.338135,-2.192915 52.338158,-2.192921 52.338181,-2.192996 52.338232,-2.193039 52.338286,-2.193042 52.338301,-2.192993 52.338516,-2.193118 52.338536,-2.193241 52.33854,-2.193279 52.338558,-2.193298 52.338583,-2.193282 52.33861,-2.193197 52.338662,-2.193114 52.338732,-2.193091 52.33878,-2.193093 52.338866,-2.193113 52.33889,-2.193146 52.338896,-2.193564 52.338869,-2.193659 52.338905,-2.193684 52.338926,-2.193687 52.338945,-2.19368 52.338959,-2.193649 52.338977,-2.193463 52.338987,-2.193412 52.339027,-2.193402 52.339068,-2.193431 52.339104,-2.19359 52.339167,-2.193627 52.339186,-2.193647 52.339209,-2.193649 52.339226,-2.193563 52.339374,-2.193567 52.339389,-2.193598 52.339433,-2.193759 52.339536,-2.193777 52.339573,-2.193777 52.339598,-2.19383 52.339621,-2.193902 52.339683,-2.193899 52.339871,-2.193917 52.339934,-2.193949 52.339997,-2.194064 52.340045,-2.19434 52.340085,-2.194694 52.34017,-2.195063 52.340284,-2.195192 52.340342,-2.195371 52.340475,-2.195538 52.340588,-2.195613 52.340654,-2.195651 52.340922,-2.195722 52.341019,-2.195756 52.341129,-2.195596 52.341258,-2.195703 52.341304,-2.196059 52.341489,-2.196331 52.34165,-2.196752 52.341869,-2.196998 52.342017,-2.197936 52.342553,-2.198575 52.342886,-2.198828 52.343041,-2.199158 52.343221,-2.200485 52.343984,-2.201049 52.344281,-2.201745 52.344605,-2.202135 52.344849,-2.202169 52.344856,-2.202238 52.344809,-2.202638 52.344986,-2.202921 52.345142,-2.203049 52.345178,-2.203315 52.345219,-2.203513 52.345199,-2.203597 52.345217,-2.203644 52.345179,-2.204124 52.345269,-2.204309 52.345315,-2.204569 52.345398,-2.20508 52.345612,-2.205099 52.345638,-2.205105 52.345683,-2.205088 52.345707,-2.20503 52.345754,-2.20502 52.345788,-2.205017 52.345892,-2.204873 52.346254,-2.204838 52.346389,-2.20481 52.346723,-2.204777 52.346866,-2.204728 52.347009,-2.204743 52.347142,-2.204744 52.347313,-2.204754 52.34744,-2.204727 52.347721,-2.204733 52.347763,-2.204728 52.348025,-2.204778 52.348823,-2.204796 52.348951,-2.204833 52.34911,-2.204961 52.349439,-2.205051 52.349579,-2.205256 52.349815,-2.205339 52.349925,-2.205385 52.350014,-2.205412 52.350099,-2.205434 52.350334,-2.205419 52.350407,-2.205373 52.350555,-2.205348 52.350594,-2.205321 52.350703,-2.205346 52.350721,-2.205909 52.350915,-2.206659 52.351241,-2.207171 52.351429,-2.20718 52.351418,-2.207192 52.351422,-2.207208 52.351407,-2.207349 52.351465,-2.207417 52.351501,-2.207661 52.35169,-2.207641 52.351705,-2.207629 52.351739,-2.207621 52.351829,-2.207586 52.35195,-2.207672 52.352092,-2.207715 52.352183,-2.207741 52.352279,-2.207849 52.352506,-2.207908 52.352587,-2.208043 52.352724,-2.208069 52.352766,-2.207994 52.352947,-2.207894 52.353235,-2.208065 52.35329,-2.208416 52.353464,-2.208818 52.353708,-2.209078 52.353852,-2.209189 52.353897,-2.210227 52.354266,-2.210374 52.354342,-2.210403 52.354321,-2.210623 52.354425,-2.210486 52.354585,-2.210409 52.35466,-2.210212 52.354876,-2.210197 52.3549,-2.210192 52.354925,-2.210201 52.354952,-2.210358 52.355064,-2.210416 52.355088,-2.210535 52.35517,-2.210669 52.35523,-2.210857 52.355344,-2.211305 52.355563,-2.211635 52.355742,-2.211844 52.355898,-2.212201 52.356073,-2.212513 52.35621,-2.212843 52.356341,-2.213949 52.356835,-2.213975 52.356844,-2.214008 52.356825,-2.214311 52.356936,-2.215069 52.357296,-2.216381 52.357974,-2.217504 52.358453,-2.218162 52.358717,-2.219002 52.359069,-2.219219 52.359138),(-2.231458 52.328493,-2.232054 52.328423,-2.232023 52.32832,-2.231973 52.328052,-2.231954 52.327913,-2.231939 52.327639,-2.231899 52.327566,-2.231832 52.326595,-2.231818 52.326589,-2.231808 52.326473,-2.231802 52.326227,-2.231808 52.326105,-2.232088 52.326066,-2.232677 52.326051,-2.232668 52.32662,-2.232656 52.326895,-2.23382 52.32688,-2.234332 52.327213,-2.23502 52.32755,-2.235316 52.327703,-2.235853 52.32781,-2.236667 52.327998,-2.23756 52.328242,-2.237491 52.328316,-2.237531 52.328369,-2.237248 52.328628,-2.23706 52.328822,-2.236899 52.329048,-2.235375 52.330839,-2.234811 52.331537,-2.234778 52.33154,-2.234783 52.331588,-2.235739 52.331666,-2.235728 52.331899,-2.235948 52.331908,-2.235949 52.331966,-2.23618 52.332005,-2.236224 52.331986,-2.236311 52.331919,-2.236357 52.331908,-2.236415 52.331903,-2.236572 52.331913,-2.236515 52.332023,-2.236433 52.332244,-2.236399 52.332387,-2.236384 52.33251,-2.236645 52.332567,-2.236622 52.332614,-2.236542 52.332722,-2.236501 52.332749,-2.236216 52.333127,-2.236017 52.33337,-2.235955 52.333468,-2.236222 52.333545,-2.235848 52.334071,-2.23581 52.334386,-2.235953 52.334372,-2.235981 52.334463,-2.236074 52.334488,-2.236107 52.334655,-2.236222 52.334657,-2.236232 52.334691,-2.23698 52.334662,-2.236809 52.33509,-2.236595 52.335514,-2.23703 52.335665,-2.236894 52.335754,-2.237107 52.335855,-2.236895 52.336012,-2.236663 52.336149,-2.236598 52.33621,-2.23677 52.336227,-2.236972 52.336268,-2.237194 52.336347,-2.237479 52.33648,-2.237425 52.336553,-2.237417 52.336636,-2.237403 52.33667,-2.23731 52.336806,-2.237095 52.337033,-2.237035 52.337013,-2.236893 52.337186,-2.236888 52.337268,-2.236822 52.337347,-2.236712 52.337538,-2.236657 52.337586,-2.236609 52.337589,-2.236503 52.337564,-2.236507 52.337541,-2.236476 52.337504,-2.235825 52.336926,-2.235453 52.336844,-2.235411 52.336882,-2.235365 52.336896,-2.235288 52.336898,-2.235195 52.336889,-2.235036 52.336861,-2.234962 52.33684,-2.234898 52.336811,-2.234796 52.336748,-2.234752 52.336709,-2.234647 52.336593,-2.234469 52.336362,-2.23438 52.336259,-2.234312 52.336192,-2.234271 52.336165,-2.234146 52.336115,-2.233681 52.336004,-2.233345 52.335865,-2.233239 52.335832,-2.233153 52.335817,-2.233175 52.335778,-2.23309 52.335749,-2.233044 52.335766,-2.232913 52.335718,-2.232751 52.335583,-2.232668 52.335498,-2.232294 52.335289,-2.2322 52.335215,-2.23205 52.33514,-2.231764 52.335029,-2.231695 52.335008,-2.231672 52.33504,-2.231469 52.334969,-2.231205 52.334899,-2.230951 52.335169,-2.230825 52.335349,-2.230803 52.335417,-2.230775 52.335453,-2.230493 52.335693,-2.230409 52.335774,-2.23023 52.335906,-2.229941 52.336199,-2.229736 52.336126,-2.229472 52.336045,-2.229283 52.336005,-2.229111 52.335979,-2.228468 52.335927,-2.227869 52.335895,-2.227471 52.335854,-2.227465 52.335808,-2.227496 52.335643,-2.227574 52.335405,-2.22768 52.335031,-2.227711 52.334799,-2.227628 52.334795,-2.227243 52.33488,-2.227151 52.334886,-2.227093 52.334882,-2.226993 52.334852,-2.226841 52.334749,-2.226608 52.334628,-2.226664 52.334481,-2.226779 52.334495,-2.226762 52.334482,-2.226799 52.334405,-2.226833 52.334393,-2.226928 52.334163,-2.226969 52.334083,-2.227039 52.33398,-2.22767 52.333978,-2.227786 52.333362,-2.227841 52.333226,-2.22794 52.333054,-2.227989 52.332986,-2.228252 52.332731,-2.229113 52.332081,-2.229285 52.331965,-2.229355 52.331924,-2.229514 52.331952,-2.230034 52.331623,-2.230079 52.331523,-2.230124 52.331494,-2.230347 52.331313,-2.231313 52.330669,-2.231657 52.330381,-2.231729 52.33031,-2.231746 52.33031,-2.231726 52.33031,-2.232135 52.329733,-2.232164 52.329681,-2.232322 52.329311,-2.23231 52.329217,-2.232267 52.329073,-2.232237 52.329015,-2.231435 52.328961,-2.231458 52.328493),(-2.180426 52.31046,-2.180616 52.310473,-2.180768 52.310497,-2.180773 52.310529,-2.180688 52.310767,-2.180648 52.310848,-2.180618 52.310989,-2.180566 52.311085,-2.180518 52.31121,-2.180435 52.311381,-2.180429 52.311395,-2.180434 52.31141,-2.181208 52.311829,-2.181117 52.311886,-2.181047 52.311916,-2.180916 52.312004,-2.180683 52.312186,-2.180416 52.312426,-2.180273 52.312569,-2.180068 52.312795,-2.179758 52.313213,-2.179442 52.313514,-2.179172 52.313748,-2.178825 52.314023,-2.178976 52.314065,-2.179431 52.314149,-2.179788 52.314225,-2.179558 52.314728,-2.179657 52.314743,-2.179666 52.314809,-2.179509 52.315053,-2.179403 52.315272,-2.179487 52.315734,-2.179438 52.315846,-2.17948 52.315922,-2.17948 52.315977,-2.179514 52.316186,-2.179523 52.316377,-2.179494 52.31642,-2.1794 52.31648,-2.179267 52.316551,-2.179197 52.316578,-2.178861 52.316542,-2.177617 52.316485,-2.177489 52.316464,-2.177335 52.316406,-2.177065 52.316278,-2.17711 52.315985,-2.177195 52.315584,-2.177379 52.315002,-2.177119 52.31494,-2.177052 52.314944,-2.17689 52.31493,-2.176608 52.314885,-2.176318 52.314825,-2.175729 52.314775,-2.175381 52.314719,-2.175182 52.314709,-2.175251 52.314403,-2.175798 52.31444,-2.175839 52.314349,-2.176204 52.314412,-2.176739 52.314487,-2.17676 52.314411,-2.176779 52.314179,-2.17658 52.314148,-2.176621 52.314077,-2.176621 52.31401,-2.176596 52.313832,-2.176882 52.313844,-2.177292 52.313811,-2.177248 52.313563,-2.177239 52.313396,-2.177257 52.313137,-2.177274 52.313034,-2.17731 52.312897,-2.177328 52.312846,-2.17734 52.312837,-2.177366 52.312744,-2.177605 52.312243,-2.177592 52.312206,-2.177617 52.312145,-2.177699 52.311988,-2.177904 52.31154,-2.177999 52.311303,-2.178016 52.311282,-2.178044 52.311271,-2.17807 52.311108,-2.178221 52.310726,-2.178243 52.31069,-2.178428 52.310117,-2.178445 52.309985,-2.178418 52.309931,-2.178419 52.309719,-2.178396 52.309566,-2.178368 52.309472,-2.178276 52.309294,-2.179224 52.309159,-2.179246 52.309399,-2.179258 52.309658,-2.179237 52.309768,-2.179213 52.309816,-2.1792 52.309976,-2.179165 52.310094,-2.179084 52.310184,-2.179022 52.310396,-2.179891 52.310464,-2.180426 52.31046),(-2.11888 52.294998,-2.118651 52.295207,-2.118479 52.295147,-2.11822 52.295079,-2.118144 52.295081,-2.118093 52.295096,-2.118001 52.295167,-2.117819 52.295079,-2.117694 52.294985,-2.117564 52.294866,-2.11732 52.294614,-2.117154 52.294467,-2.116923 52.294202,-2.116576 52.29387,-2.115397 52.292876,-2.115299 52.292939,-2.115354 52.292989,-2.114965 52.293248,-2.114625 52.293051,-2.114526 52.293092,-2.114357 52.293205,-2.113672 52.293703,-2.113582 52.293832,-2.113345 52.294,-2.11324 52.294103,-2.113055 52.294314,-2.112694 52.294206,-2.113142 52.293841,-2.113204 52.293782,-2.11293 52.293676,-2.112789 52.293609,-2.112455 52.293469,-2.112367 52.293448,-2.112255 52.293404,-2.112081 52.293324,-2.111216 52.292859,-2.110677 52.292517,-2.109852 52.292111,-2.10979 52.292062,-2.109674 52.291991,-2.109091 52.291728,-2.108843 52.291629,-2.108674 52.291574,-2.1085 52.291485,-2.107887 52.291202,-2.107546 52.29107,-2.107304 52.290998,-2.107161 52.290966,-2.107 52.290941,-2.106825 52.290924,-2.106447 52.290915,-2.106122 52.290895,-2.105633 52.290837,-2.105667 52.29069,-2.105664 52.290592,-2.105717 52.290494,-2.105798 52.290413,-2.105916 52.290376,-2.10617 52.290273,-2.106526 52.290081,-2.106783 52.289967,-2.107428 52.289556,-2.107703 52.289405,-2.108111 52.289155,-2.10836 52.289018,-2.108549 52.288932,-2.108689 52.288859,-2.108749 52.288818,-2.109105 52.28864,-2.109436 52.28845,-2.109756 52.288248,-2.109926 52.288163,-2.110305 52.287948,-2.110537 52.287829,-2.110759 52.287729,-2.110976 52.28765,-2.11124 52.287546,-2.111354 52.287491,-2.111522 52.287432,-2.112074 52.28721,-2.11243 52.287044,-2.11309 52.28668,-2.11327 52.286602,-2.113539 52.286509,-2.113903 52.286439,-2.114339 52.286368,-2.114722 52.286346,-2.11512 52.286367,-2.115435 52.286399,-2.115567 52.286426,-2.115797 52.286491,-2.115909 52.286547,-2.116039 52.28667,-2.116077 52.286733,-2.116109 52.286827,-2.116136 52.286864,-2.116164 52.286946,-2.116186 52.286979,-2.116289 52.287201,-2.11642 52.287336,-2.116505 52.287398,-2.11663 52.28747,-2.116847 52.287556,-2.117023 52.287593,-2.117335 52.28764,-2.117442 52.287673,-2.117467 52.2877,-2.11753 52.28781,-2.117597 52.287896,-2.117855 52.288115,-2.118489 52.288625,-2.118613 52.288748,-2.119051 52.288733,-2.119252 52.288705,-2.119446 52.288667,-2.11957 52.288689,-2.119711 52.288737,-2.119817 52.288791,-2.119896 52.288845,-2.11994 52.288894,-2.120117 52.289131,-2.120188 52.289282,-2.120259 52.289465,-2.12041 52.289965,-2.12058 52.290662,-2.120615 52.290873,-2.120642 52.291148,-2.120635 52.291613,-2.120591 52.291826,-2.120576 52.291822,-2.120518 52.292196,-2.120445 52.292538,-2.120191 52.293109,-2.120018 52.29346,-2.119973 52.293531,-2.119588 52.294126,-2.119308 52.294513,-2.119156 52.294717,-2.11888 52.294998)))"  # noqa: E501
-    expected = "MULTIPOLYGON (((-2.235523 52.360881,-2.237210 52.360857,-2.237512 52.360842,-2.237998 52.360783,-2.239279 52.360553,-2.240296 52.360428,-2.241611 52.360336,-2.241935 52.360368,-2.242738 52.360753,-2.242828 52.360787,-2.242966 52.360819,-2.244501 52.360918,-2.245688 52.361022,-2.246176 52.360063,-2.246586 52.359370,-2.246946 52.358795,-2.247101 52.358432,-2.246931 52.358449,-2.247499 52.357102,-2.247668 52.356678,-2.247808 52.356430,-2.247931 52.356144,-2.248250 52.355477,-2.248235 52.355473,-2.248248 52.355445,-2.247603 52.355271,-2.247506 52.355255,-2.247419 52.355261,-2.247287 52.355305,-2.245728 52.354927,-2.245735 52.354917,-2.245707 52.354910,-2.246763 52.353241,-2.245820 52.353017,-2.246116 52.352627,-2.247583 52.350831,-2.247690 52.350784,-2.249376 52.348833,-2.249439 52.348856,-2.249448 52.348846,-2.249802 52.348973,-2.249878 52.348847,-2.250434 52.348079,-2.250952 52.347400,-2.251079 52.347203,-2.251191 52.347056,-2.251216 52.347010,-2.251659 52.345845,-2.251429 52.345849,-2.251097 52.345840,-2.250744 52.345802,-2.249882 52.345656,-2.249750 52.345653,-2.249516 52.345665,-2.249537 52.345640,-2.249415 52.345652,-2.249481 52.345606,-2.249765 52.345473,-2.250106 52.345334,-2.251117 52.344948,-2.251371 52.344825,-2.251439 52.344783,-2.251750 52.344550,-2.252118 52.344229,-2.252137 52.344207,-2.252234 52.344028,-2.252596 52.343282,-2.252628 52.343203,-2.252754 52.342968,-2.252872 52.342615,-2.253145 52.342118,-2.253206 52.342017,-2.253333 52.341845,-2.253932 52.341125,-2.253981 52.341142,-2.254006 52.341115,-2.254735 52.341322,-2.255745 52.341556,-2.256257 52.341633,-2.256417 52.341647,-2.256684 52.341653,-2.256809 52.341679,-2.257019 52.341683,-2.257338 52.341658,-2.257543 52.341623,-2.257789 52.341559,-2.257946 52.341506,-2.260411 52.341036,-2.261065 52.340953,-2.261594 52.340866,-2.261983 52.340815,-2.262009 52.340743,-2.261232 52.340654,-2.261227 52.340648,-2.261243 52.340535,-2.261266 52.340461,-2.261301 52.340388,-2.261386 52.340260,-2.261700 52.339902,-2.261730 52.339830,-2.261737 52.339784,-2.261737 52.339728,-2.261725 52.339670,-2.261676 52.339575,-2.261611 52.339504,-2.261219 52.339217,-2.261089 52.339141,-2.260976 52.339101,-2.260919 52.339070,-2.260848 52.339005,-2.260709 52.338646,-2.260736 52.338406,-2.260872 52.338325,-2.261047 52.338248,-2.261181 52.338204,-2.261798 52.338058,-2.263100 52.338127,-2.263077 52.338016,-2.263068 52.337922,-2.263072 52.337830,-2.263285 52.336903,-2.263308 52.336841,-2.263370 52.336718,-2.263858 52.336039,-2.264079 52.335934,-2.264950 52.335971,-2.265184 52.335971,-2.265538 52.335952,-2.265868 52.335904,-2.265982 52.335899,-2.266097 52.335896,-2.266333 52.335924,-2.266826 52.336150,-2.267131 52.336275,-2.268065 52.336632,-2.268295 52.336425,-2.268478 52.336282,-2.268340 52.336212,-2.268366 52.336176,-2.268478 52.336077,-2.268753 52.336065,-2.268876 52.336092,-2.268998 52.336126,-2.269216 52.336269,-2.269160 52.336135,-2.269161 52.336057,-2.269175 52.336000,-2.269208 52.335931,-2.269235 52.335896,-2.269305 52.335841,-2.269229 52.335752,-2.269216 52.335710,-2.269213 52.335656,-2.269220 52.335626,-2.269269 52.335550,-2.269719 52.335152,-2.267975 52.334078,-2.267306 52.334160,-2.266837 52.333635,-2.267003 52.333569,-2.267208 52.333507,-2.267342 52.333512,-2.267833 52.333485,-2.267968 52.333740,-2.270334 52.333121,-2.271779 52.332618,-2.271123 52.331963,-2.270858 52.331634,-2.270783 52.331658,-2.270828 52.331784,-2.270820 52.331804,-2.270767 52.331862,-2.270514 52.332056,-2.270431 52.332094,-2.270346 52.332115,-2.270170 52.332080,-2.269903 52.331995,-2.269734 52.331928,-2.269379 52.331838,-2.268946 52.331646,-2.268627 52.331472,-2.268517 52.331380,-2.268476 52.331326,-2.268468 52.331296,-2.268376 52.331228,-2.267261 52.330812,-2.266464 52.330537,-2.266292 52.330461,-2.266156 52.330390,-2.265971 52.330276,-2.265553 52.330117,-2.265334 52.330051,-2.264944 52.329959,-2.264135 52.329802,-2.263841 52.329757,-2.263447 52.329637,-2.262763 52.329639,-2.262622 52.329625,-2.261951 52.329446,-2.261659 52.329297,-2.261422 52.329153,-2.261161 52.328932,-2.260899 52.328771,-2.260250 52.328586,-2.259985 52.328528,-2.258908 52.328352,-2.258338 52.328205,-2.258204 52.328165,-2.257965 52.328072,-2.257672 52.327935,-2.257207 52.327676,-2.256751 52.327382,-2.256239 52.327071,-2.255899 52.326878,-2.255135 52.326500,-2.254356 52.326155,-2.253899 52.325995,-2.251895 52.325501,-2.251610 52.325417,-2.251494 52.325375,-2.251266 52.325266,-2.251065 52.325164,-2.249734 52.324428,-2.249612 52.324343,-2.249415 52.324182,-2.249009 52.323791,-2.248500 52.323317,-2.248470 52.323311,-2.248384 52.323328,-2.248302 52.323356,-2.247721 52.323621,-2.246956 52.323937,-2.246810 52.323980,-2.246701 52.323996,-2.246417 52.324011,-2.246082 52.324003,-2.244877 52.323844,-2.244618 52.323833,-2.243990 52.323888,-2.243637 52.323931,-2.242906 52.324036,-2.242683 52.324085,-2.242216 52.324138,-2.242029 52.324137,-2.241791 52.324096,-2.241159 52.324009,-2.240891 52.323975,-2.240744 52.323970,-2.240629 52.323990,-2.240582 52.324014,-2.240491 52.324088,-2.240427 52.324129,-2.240371 52.324150,-2.240193 52.324187,-2.240086 52.324194,-2.238681 52.324191,-2.238192 52.324197,-2.238066 52.324189,-2.237960 52.324173,-2.237192 52.324008,-2.237048 52.323984,-2.236532 52.323930,-2.236276 52.323914,-2.235472 52.323893,-2.235268 52.323905,-2.234935 52.323939,-2.233882 52.324114,-2.233390 52.324157,-2.233299 52.324152,-2.233133 52.324127,-2.232933 52.324082,-2.232680 52.324010,-2.232561 52.324002,-2.232418 52.324018,-2.231378 52.324023,-2.230386 52.323910,-2.229426 52.323860,-2.229168 52.323839,-2.228476 52.323780,-2.227666 52.323688,-2.226645 52.323599,-2.225510 52.323555,-2.224674 52.323548,-2.222433 52.323576,-2.221770 52.323593,-2.219554 52.323676,-2.219186 52.323678,-2.218596 52.323663,-2.217935 52.323615,-2.217612 52.323579,-2.215772 52.323310,-2.215379 52.323260,-2.214933 52.323219,-2.214828 52.323091,-2.214618 52.322731,-2.214195 52.322138,-2.213858 52.321599,-2.212909 52.320136,-2.212881 52.320139,-2.212870 52.320133,-2.212865 52.320115,-2.212891 52.320104,-2.212668 52.319760,-2.211414 52.318001,-2.211164 52.317585,-2.210969 52.317285,-2.210555 52.316763,-2.210414 52.316539,-2.210150 52.316157,-2.209525 52.315315,-2.209492 52.315250,-2.209155 52.314907,-2.208656 52.314437,-2.208045 52.313948,-2.207645 52.313607,-2.207454 52.313477,-2.207151 52.313302,-2.206783 52.313110,-2.205236 52.312387,-2.204279 52.312002,-2.204041 52.311918,-2.204048 52.311909,-2.203804 52.311811,-2.202980 52.311505,-2.200164 52.310598,-2.199682 52.310487,-2.199274 52.310432,-2.198991 52.310363,-2.198039 52.310094,-2.197348 52.309915,-2.197041 52.309843,-2.195562 52.309458,-2.194698 52.309179,-2.194607 52.309171,-2.194566 52.309192,-2.194059 52.308965,-2.193892 52.308899,-2.193484 52.308773,-2.192730 52.308559,-2.192514 52.308476,-2.191466 52.308154,-2.190573 52.307851,-2.190030 52.307644,-2.189245 52.307362,-2.187706 52.306743,-2.187383 52.306630,-2.186503 52.306229,-2.186078 52.306025,-2.185728 52.305831,-2.185045 52.305505,-2.184750 52.305374,-2.184313 52.305215,-2.183867 52.305034,-2.183517 52.304835,-2.183496 52.304815,-2.183484 52.304788,-2.183208 52.304683,-2.182763 52.304487,-2.182488 52.304350,-2.182340 52.304298,-2.182240 52.304245,-2.182218 52.304224,-2.181076 52.303659,-2.180490 52.303383,-2.180005 52.303049,-2.179765 52.302866,-2.179494 52.302703,-2.178980 52.302352,-2.178363 52.301873,-2.177925 52.301577,-2.177816 52.301495,-2.177507 52.301243,-2.177305 52.301043,-2.177306 52.301030,-2.177284 52.300980,-2.177306 52.300952,-2.177323 52.300943,-2.177309 52.300924,-2.177062 52.300721,-2.176974 52.300670,-2.176712 52.300431,-2.176301 52.300013,-2.176175 52.299895,-2.175687 52.299377,-2.175661 52.299366,-2.175549 52.299267,-2.175285 52.299094,-2.173739 52.297686,-2.172665 52.296728,-2.172306 52.296391,-2.172212 52.296285,-2.172144 52.296189,-2.171884 52.295898,-2.170878 52.294971,-2.170810 52.294911,-2.170793 52.294916,-2.169356 52.293535,-2.168710 52.292801,-2.168576 52.292688,-2.167765 52.291862,-2.167285 52.291402,-2.166910 52.291070,-2.166472 52.290739,-2.166227 52.290542,-2.165783 52.290218,-2.165184 52.289735,-2.164453 52.288998,-2.164423 52.288935,-2.164337 52.288937,-2.164137 52.288771,-2.163589 52.288274,-2.163177 52.287867,-2.162777 52.287424,-2.162721 52.287374,-2.162832 52.287344,-2.162840 52.287323,-2.162815 52.287287,-2.162490 52.286994,-2.162381 52.286936,-2.162309 52.286888,-2.161770 52.286354,-2.161519 52.286127,-2.161303 52.285947,-2.161207 52.285869,-2.161132 52.285890,-2.161071 52.285863,-2.160828 52.285572,-2.160580 52.285254,-2.160223 52.284829,-2.160187 52.284799,-2.160082 52.284810,-2.160053 52.284772,-2.160069 52.284767,-2.160057 52.284745,-2.160098 52.284696,-2.159613 52.284357,-2.159406 52.284252,-2.159208 52.284046,-2.159238 52.284030,-2.158606 52.283540,-2.158363 52.283287,-2.158090 52.283076,-2.157663 52.282713,-2.157522 52.282580,-2.156648 52.281699,-2.155863 52.280815,-2.155793 52.280761,-2.155724 52.280694,-2.155056 52.280027,-2.154887 52.279811,-2.154507 52.279234,-2.154260 52.278803,-2.153952 52.278162,-2.153788 52.277774,-2.153570 52.277205,-2.153479 52.276740,-2.153344 52.275658,-2.153337 52.275562,-2.153342 52.275076,-2.153362 52.274834,-2.153360 52.274723,-2.153328 52.274689,-2.153340 52.274618,-2.153118 52.274561,-2.153040 52.274629,-2.152931 52.274662,-2.152261 52.274542,-2.151990 52.274509,-2.151777 52.274472,-2.151575 52.274421,-2.150223 52.274027,-2.150059 52.273941,-2.149812 52.273850,-2.149374 52.273737,-2.148982 52.273623,-2.148098 52.273385,-2.147926 52.273365,-2.147052 52.273224,-2.146156 52.273065,-2.145637 52.273027,-2.145325 52.273023,-2.145044 52.273049,-2.144518 52.273107,-2.144083 52.273197,-2.143846 52.273233,-2.142873 52.273311,-2.142723 52.273330,-2.140601 52.273627,-2.140123 52.273737,-2.139879 52.273321,-2.139849 52.273288,-2.139793 52.273089,-2.139801 52.273055,-2.139758 52.272902,-2.139874 52.272858,-2.139885 52.272842,-2.140452 52.272424,-2.140819 52.272123,-2.141517 52.271585,-2.140963 52.271307,-2.140668 52.271093,-2.140684 52.270594,-2.140745 52.270590,-2.140742 52.270541,-2.140825 52.269835,-2.141008 52.269702,-2.141013 52.269601,-2.140619 52.269575,-2.140299 52.269534,-2.139733 52.269483,-2.139028 52.269498,-2.138102 52.269489,-2.137727 52.269507,-2.137247 52.269544,-2.137088 52.269546,-2.136636 52.269532,-2.136455 52.269537,-2.135216 52.269605,-2.135134 52.269630,-2.135093 52.269673,-2.135086 52.269712,-2.135095 52.269774,-2.135035 52.269782,-2.135001 52.269717,-2.134906 52.269681,-2.134790 52.269683,-2.134423 52.269743,-2.134042 52.269837,-2.133267 52.269965,-2.132411 52.270121,-2.131910 52.270197,-2.131867 52.270103,-2.131861 52.270066,-2.131868 52.270052,-2.131906 52.270011,-2.131973 52.269966,-2.132107 52.269919,-2.132177 52.269857,-2.132205 52.269808,-2.132210 52.269778,-2.132204 52.269692,-2.132210 52.269664,-2.132194 52.269627,-2.132144 52.269583,-2.132144 52.269490,-2.132124 52.269451,-2.132132 52.269426,-2.132174 52.269395,-2.132297 52.269370,-2.132344 52.269353,-2.132385 52.269296,-2.132458 52.269243,-2.132819 52.269195,-2.132635 52.268913,-2.132521 52.268863,-2.131987 52.268576,-2.131654 52.268386,-2.132247 52.268106,-2.132677 52.267881,-2.133159 52.267691,-2.133525 52.267524,-2.133697 52.267436,-2.133681 52.267397,-2.133704 52.267370,-2.133801 52.267356,-2.134080 52.267210,-2.134263 52.267096,-2.134357 52.267027,-2.134536 52.266872,-2.134395 52.266809,-2.133882 52.266634,-2.133596 52.266567,-2.133274 52.266519,-2.132821 52.266488,-2.131841 52.266480,-2.131317 52.266456,-2.130986 52.266432,-2.130270 52.266281,-2.129602 52.266161,-2.129198 52.266117,-2.128694 52.266075,-2.128091 52.266053,-2.127364 52.266075,-2.126837 52.266120,-2.126522 52.266128,-2.126212 52.266119,-2.119952 52.265298,-2.115125 52.264694,-2.111480 52.264109,-2.111087 52.264058,-2.110021 52.263900,-2.109683 52.263867,-2.108934 52.263822,-2.107455 52.263701,-2.106395 52.263598,-2.105909 52.263562,-2.103320 52.263314,-2.101655 52.263172,-2.100815 52.263090,-2.097650 52.262809,-2.096722 52.262699,-2.093677 52.262226,-2.091837 52.262005,-2.090728 52.261882,-2.089397 52.261757,-2.086862 52.261558,-2.086106 52.261448,-2.083137 52.261193,-2.080816 52.260884,-2.077210 52.260601,-2.072849 52.260144,-2.071848 52.260094,-2.071465 52.260083,-2.070513 52.260085,-2.070040 52.260097,-2.069706 52.260110,-2.068625 52.260191,-2.067797 52.260290,-2.067344 52.260359,-2.066804 52.260453,-2.065932 52.260633,-2.061021 52.261689,-2.060301 52.261871,-2.059602 52.262071,-2.057190 52.262856,-2.055880 52.263301,-2.055105 52.263620,-2.054654 52.263835,-2.054334 52.264000,-2.052650 52.264933,-2.054113 52.265903,-2.053168 52.266439,-2.053286 52.266516,-2.053487 52.266625,-2.053178 52.266876,-2.053024 52.266979,-2.052999 52.267033,-2.053277 52.267229,-2.053316 52.267264,-2.053050 52.267393,-2.053179 52.267468,-2.052755 52.267682,-2.053013 52.268006,-2.052842 52.268069,-2.052624 52.268163,-2.051959 52.268357,-2.051412 52.268538,-2.051342 52.268633,-2.051278 52.268704,-2.051134 52.268926,-2.050976 52.268815,-2.050820 52.268675,-2.050594 52.268437,-2.050524 52.268344,-2.050376 52.268077,-2.050344 52.267990,-2.050344 52.267943,-2.050273 52.267892,-2.050228 52.267839,-2.050292 52.267799,-2.050331 52.267732,-2.050466 52.267552,-2.050800 52.267331,-2.050845 52.267284,-2.050925 52.266893,-2.050903 52.266893,-2.051031 52.266529,-2.051008 52.266523,-2.051053 52.266391,-2.051079 52.266281,-2.051012 52.266034,-2.050335 52.265812,-2.049705 52.265702,-2.049912 52.265300,-2.049183 52.265186,-2.048167 52.264985,-2.048042 52.264947,-2.047824 52.264865,-2.047609 52.264775,-2.047040 52.264509,-2.046443 52.264280,-2.046032 52.264164,-2.045780 52.264114,-2.045098 52.263958,-2.044804 52.263875,-2.044171 52.263675,-2.043079 52.263288,-2.041948 52.262863,-2.040945 52.262503,-2.040575 52.262352,-2.040183 52.262177,-2.040023 52.262067,-2.039769 52.261928,-2.039461 52.261783,-2.038694 52.261470,-2.036469 52.260611,-2.035455 52.260186,-2.032303 52.258986,-2.030577 52.258311,-2.027489 52.257155,-2.026896 52.256947,-2.026198 52.256724,-2.025620 52.256556,-2.025304 52.256473,-2.024781 52.256343,-2.022064 52.255722,-2.020570 52.255342,-2.020223 52.255241,-2.019520 52.254997,-2.019093 52.254829,-2.018639 52.254628,-2.017926 52.254306,-2.017139 52.253933,-2.015910 52.253315,-2.014882 52.252767,-2.014409 52.253113,-2.014356 52.253125,-2.014104 52.253149,-2.013732 52.253224,-2.013712 52.253322,-2.013741 52.253391,-2.013744 52.253416,-2.013713 52.253546,-2.013652 52.253690,-2.013634 52.253755,-2.013501 52.254050,-2.013463 52.254048,-2.013403 52.254201,-2.013282 52.254588,-2.013248 52.254836,-2.013153 52.255160,-2.013133 52.255403,-2.013234 52.255501,-2.013272 52.255502,-2.013348 52.255523,-2.013531 52.255657,-2.013751 52.256016,-2.013817 52.256170,-2.013782 52.256480,-2.013789 52.256555,-2.013582 52.257011,-2.013461 52.257150,-2.013359 52.257330,-2.013034 52.257735,-2.013340 52.257619,-2.013313 52.257656,-2.013385 52.257629,-2.013338 52.257680,-2.013151 52.257941,-2.013092 52.258006,-2.013095 52.258002,-2.013036 52.258000,-2.012890 52.257954,-2.012852 52.257965,-2.012765 52.258104,-2.012738 52.258224,-2.012724 52.258248,-2.012591 52.258394,-2.012600 52.258479,-2.012552 52.258562,-2.012559 52.258645,-2.012410 52.258817,-2.012306 52.258914,-2.012231 52.258966,-2.012228 52.258983,-2.012253 52.259027,-2.012251 52.259043,-2.012234 52.259088,-2.012133 52.259274,-2.011988 52.259383,-2.011896 52.259463,-2.011451 52.259926,-2.011405 52.259997,-2.011409 52.260046,-2.011453 52.260145,-2.011436 52.260186,-2.011425 52.260246,-2.011444 52.260498,-2.011480 52.260613,-2.011548 52.260770,-2.011510 52.260964,-2.011516 52.261016,-2.011506 52.261126,-2.011521 52.261154,-2.011553 52.261272,-2.011532 52.261312,-2.011534 52.261374,-2.011552 52.261594,-2.011579 52.261796,-2.011581 52.261883,-2.011565 52.262038,-2.011526 52.262265,-2.011487 52.262351,-2.011412 52.262482,-2.011280 52.262589,-2.011262 52.262614,-2.011249 52.262674,-2.011248 52.262794,-2.011276 52.262881,-2.011426 52.262932,-2.011869 52.262983,-2.012100 52.263027,-2.012183 52.263056,-2.012175 52.263235,-2.012164 52.263313,-2.012164 52.263431,-2.012205 52.263667,-2.012152 52.263722,-2.012147 52.263746,-2.012163 52.263892,-2.012217 52.264040,-2.012227 52.264166,-2.012260 52.264297,-2.012264 52.264371,-2.012275 52.264403,-2.012591 52.264325,-2.013001 52.264177,-2.013054 52.264169,-2.013158 52.264179,-2.013187 52.264193,-2.013220 52.264214,-2.013286 52.264300,-2.013296 52.264519,-2.013334 52.264649,-2.013388 52.264720,-2.013477 52.264903,-2.013668 52.265134,-2.013720 52.265216,-2.013916 52.265420,-2.013923 52.265454,-2.013918 52.265566,-2.013993 52.265666,-2.013980 52.265676,-2.014000 52.265700,-2.013893 52.265737,-2.013819 52.265789,-2.013783 52.265824,-2.013514 52.265910,-2.013339 52.265986,-2.013267 52.266036,-2.013246 52.266113,-2.013482 52.266265,-2.014269 52.266514,-2.014445 52.266553,-2.014498 52.266558,-2.014603 52.266586,-2.014808 52.266687,-2.014899 52.266748,-2.015227 52.267016,-2.015368 52.267141,-2.015425 52.267207,-2.015453 52.267277,-2.015451 52.267320,-2.015388 52.267461,-2.015382 52.267660,-2.015391 52.267671,-2.015378 52.267725,-2.015383 52.267764,-2.015418 52.267838,-2.015395 52.267873,-2.015410 52.268023,-2.015456 52.268177,-2.015443 52.268264,-2.015526 52.268443,-2.015562 52.268655,-2.015519 52.268767,-2.015499 52.269305,-2.015553 52.269538,-2.015538 52.269609,-2.015556 52.269713,-2.015550 52.269962,-2.015538 52.270010,-2.015300 52.270031,-2.014998 52.270037,-2.014797 52.270031,-2.014631 52.270046,-2.014363 52.270089,-2.014171 52.270104,-2.014101 52.270119,-2.013838 52.270183,-2.013746 52.270228,-2.013726 52.270357,-2.013701 52.270446,-2.013656 52.270462,-2.013412 52.270480,-2.013175 52.270473,-2.012945 52.270455,-2.012680 52.270420,-2.012137 52.270535,-2.012061 52.270569,-2.011535 52.270887,-2.011513 52.270908,-2.011293 52.270900,-2.011173 52.270907,-2.010471 52.270896,-2.009828 52.270896,-2.009388 52.270761,-2.008923 52.270602,-2.008777 52.270568,-2.008708 52.270523,-2.008582 52.270486,-2.008516 52.270485,-2.008015 52.270407,-2.007955 52.270408,-2.007854 52.270424,-2.007763 52.270422,-2.007177 52.270265,-2.007111 52.270269,-2.006789 52.270324,-2.006608 52.270378,-2.006566 52.270406,-2.006532 52.270450,-2.006443 52.270494,-2.006148 52.270706,-2.006078 52.270737,-2.005848 52.270595,-2.005266 52.271090,-2.004388 52.271857,-2.004221 52.272012,-2.003228 52.273624,-2.003955 52.273842,-2.004791 52.274226,-2.005690 52.274575,-2.005878 52.274660,-2.005977 52.274711,-2.006112 52.274797,-2.006391 52.275022,-2.006553 52.275136,-2.007052 52.275373,-2.007026 52.275451,-2.006992 52.275478,-2.006867 52.275564,-2.006796 52.275602,-2.006716 52.275662,-2.006684 52.275663,-2.006665 52.275673,-2.006399 52.275670,-2.006143 52.275652,-2.005830 52.275642,-2.005412 52.275662,-2.004867 52.275877,-2.003539 52.276423,-2.003186 52.276559,-2.002500 52.276792,-2.002582 52.276892,-2.002702 52.277066,-2.002837 52.277292,-2.002938 52.277488,-2.003110 52.277767,-2.003336 52.278010,-2.003635 52.278297,-2.003917 52.278591,-2.003909 52.278597,-2.003919 52.278608,-2.003466 52.278914,-2.002685 52.279416,-2.002505 52.279545,-2.002665 52.279522,-2.002815 52.279485,-2.003018 52.279448,-2.003227 52.279393,-2.003568 52.279320,-2.003733 52.279257,-2.003921 52.279229,-2.003987 52.279247,-2.004053 52.279305,-2.004126 52.279349,-2.004596 52.279500,-2.004716 52.279510,-2.004758 52.279494,-2.004825 52.279436,-2.005009 52.279218,-2.005023 52.279182,-2.005025 52.279114,-2.005007 52.279066,-2.005708 52.278795,-2.006203 52.278545,-2.006307 52.278485,-2.006395 52.278490,-2.006579 52.278526,-2.006643 52.278527,-2.006733 52.278553,-2.007261 52.278589,-2.007947 52.278518,-2.007983 52.278578,-2.007989 52.278618,-2.008017 52.278654,-2.008122 52.278600,-2.008213 52.278569,-2.008370 52.278477,-2.008906 52.278106,-2.009038 52.278075,-2.009076 52.278056,-2.009161 52.278046,-2.009232 52.278059,-2.009298 52.278084,-2.009384 52.278138,-2.009466 52.278171,-2.009536 52.278215,-2.009560 52.278221,-2.009602 52.278257,-2.009929 52.278460,-2.010008 52.278489,-2.010053 52.278517,-2.010259 52.278665,-2.010362 52.278729,-2.010930 52.278540,-2.011356 52.278422,-2.011701 52.278353,-2.011758 52.278355,-2.011893 52.278383,-2.012173 52.278412,-2.012446 52.278425,-2.013610 52.278473,-2.014380 52.278494,-2.014757 52.278519,-2.014887 52.278492,-2.015058 52.278502,-2.015488 52.278567,-2.015719 52.278582,-2.016674 52.278770,-2.016983 52.278820,-2.017044 52.278812,-2.017095 52.278816,-2.017218 52.278855,-2.017597 52.278908,-2.017876 52.278989,-2.018143 52.279011,-2.018843 52.279148,-2.018915 52.279164,-2.019167 52.279254,-2.019332 52.279303,-2.019723 52.279407,-2.020041 52.279527,-2.020230 52.279396,-2.020339 52.279157,-2.020442 52.279098,-2.020473 52.279097,-2.020564 52.279035,-2.020644 52.278855,-2.020841 52.278755,-2.020919 52.278664,-2.020937 52.278666,-2.020994 52.278719,-2.021047 52.278732,-2.021128 52.278700,-2.021161 52.278660,-2.021196 52.278644,-2.021301 52.278621,-2.021382 52.278565,-2.021611 52.278498,-2.021806 52.278490,-2.022201 52.278402,-2.022338 52.278355,-2.022439 52.278336,-2.022470 52.278316,-2.022581 52.278325,-2.022653 52.278307,-2.022738 52.278363,-2.022885 52.278411,-2.022910 52.278408,-2.023106 52.278325,-2.023263 52.278086,-2.023281 52.278113,-2.023313 52.278054,-2.023329 52.278047,-2.023387 52.278197,-2.023396 52.278284,-2.023461 52.278389,-2.023499 52.278400,-2.023599 52.278390,-2.023628 52.278408,-2.023625 52.278473,-2.023613 52.278497,-2.023587 52.278521,-2.023531 52.278549,-2.023433 52.278577,-2.023421 52.278642,-2.023618 52.278658,-2.023660 52.278656,-2.023742 52.278634,-2.023813 52.278631,-2.023826 52.278634,-2.023836 52.278655,-2.023825 52.278679,-2.023826 52.278699,-2.023918 52.278730,-2.023985 52.278720,-2.024014 52.278733,-2.024014 52.278762,-2.023970 52.278819,-2.023951 52.278832,-2.023897 52.278834,-2.023885 52.278846,-2.023917 52.278925,-2.023964 52.278989,-2.023957 52.279030,-2.023965 52.279057,-2.024024 52.279158,-2.024026 52.279195,-2.024008 52.279231,-2.024014 52.279279,-2.023961 52.279298,-2.023959 52.279319,-2.023992 52.279443,-2.024086 52.279561,-2.024209 52.279544,-2.024224 52.279567,-2.024263 52.279563,-2.024282 52.279579,-2.024296 52.279612,-2.024296 52.279671,-2.024400 52.279734,-2.024476 52.279822,-2.024447 52.279860,-2.024444 52.279878,-2.024457 52.279889,-2.024497 52.279896,-2.024521 52.279909,-2.024516 52.279989,-2.024560 52.279980,-2.024615 52.279987,-2.024705 52.280015,-2.024705 52.280035,-2.024664 52.280052,-2.024640 52.280078,-2.024678 52.280113,-2.024680 52.280126,-2.024743 52.280172,-2.024799 52.280183,-2.024835 52.280202,-2.024840 52.280218,-2.024827 52.280298,-2.024832 52.280368,-2.024856 52.280376,-2.025076 52.280357,-2.025243 52.280330,-2.025950 52.280616,-2.026153 52.280647,-2.026219 52.280641,-2.026310 52.280712,-2.026505 52.280822,-2.026628 52.280960,-2.026869 52.280978,-2.027077 52.280958,-2.027300 52.280952,-2.027480 52.280960,-2.027622 52.280994,-2.027729 52.281031,-2.027948 52.281142,-2.028159 52.281188,-2.028220 52.281275,-2.028297 52.281338,-2.028338 52.281356,-2.028407 52.281356,-2.028442 52.281365,-2.028578 52.281491,-2.028646 52.281530,-2.028874 52.281633,-2.029714 52.282096,-2.030078 52.282355,-2.030281 52.282461,-2.030501 52.282534,-2.030618 52.282552,-2.030967 52.282554,-2.031084 52.282538,-2.031442 52.282521,-2.031706 52.282521,-2.031773 52.282526,-2.031840 52.282549,-2.032201 52.282622,-2.032623 52.282664,-2.032788 52.282637,-2.033123 52.282637,-2.033135 52.282663,-2.033159 52.282664,-2.033206 52.282747,-2.033229 52.282837,-2.033331 52.282878,-2.033503 52.282831,-2.033858 52.282644,-2.033854 52.282634,-2.033868 52.282638,-2.033895 52.282624,-2.034407 52.282818,-2.034377 52.282837,-2.034396 52.282850,-2.034087 52.283053,-2.033945 52.283161,-2.034116 52.283238,-2.034345 52.283292,-2.034774 52.283439,-2.034812 52.283467,-2.035173 52.283597,-2.035322 52.283676,-2.035409 52.283757,-2.035388 52.283766,-2.035461 52.283847,-2.035463 52.283868,-2.035482 52.283898,-2.035579 52.284148,-2.035655 52.284296,-2.035982 52.284204,-2.035995 52.284212,-2.036023 52.284203,-2.036092 52.284259,-2.037230 52.284910,-2.037793 52.285224,-2.037847 52.285230,-2.037851 52.285241,-2.038491 52.285625,-2.039867 52.286008,-2.040465 52.286181,-2.040725 52.286264,-2.041175 52.286364,-2.041472 52.286455,-2.041636 52.286520,-2.042654 52.287070,-2.043018 52.287244,-2.043205 52.287312,-2.043435 52.287379,-2.044189 52.287580,-2.044492 52.287647,-2.045021 52.287711,-2.045283 52.287734,-2.045534 52.287771,-2.046079 52.287795,-2.046641 52.287796,-2.046863 52.287806,-2.047014 52.287823,-2.047079 52.287838,-2.047852 52.287700,-2.047892 52.287640,-2.048324 52.287533,-2.048860 52.287309,-2.048904 52.287313,-2.048950 52.287295,-2.048958 52.287316,-2.048966 52.287316,-2.049629 52.287195,-2.050866 52.287057,-2.051209 52.286957,-2.052545 52.286378,-2.052929 52.286177,-2.053228 52.285970,-2.053522 52.285785,-2.053954 52.285492,-2.054364 52.285265,-2.054472 52.285171,-2.054576 52.285133,-2.054587 52.285142,-2.055054 52.284908,-2.055490 52.284742,-2.055516 52.284719,-2.055515 52.284704,-2.055550 52.284690,-2.056196 52.284552,-2.056796 52.284440,-2.057510 52.284337,-2.057668 52.284298,-2.057820 52.284279,-2.058061 52.284275,-2.058129 52.284282,-2.058266 52.284314,-2.058447 52.284384,-2.058609 52.284383,-2.059424 52.284297,-2.059726 52.284239,-2.059722 52.284232,-2.060016 52.284168,-2.060405 52.284103,-2.060679 52.284171,-2.060850 52.284176,-2.060935 52.284155,-2.061002 52.284219,-2.061398 52.284438,-2.061905 52.284742,-2.062286 52.284864,-2.062883 52.284939,-2.063442 52.285044,-2.063825 52.285177,-2.064109 52.285260,-2.067122 52.286288,-2.067270 52.286325,-2.067770 52.286416,-2.068365 52.286557,-2.070442 52.287148,-2.070643 52.287224,-2.070829 52.287278,-2.071135 52.287398,-2.071137 52.287417,-2.071155 52.287426,-2.071139 52.287446,-2.071147 52.287524,-2.071307 52.287637,-2.071509 52.287721,-2.071695 52.287777,-2.071703 52.287769,-2.073369 52.288352,-2.074205 52.288621,-2.074508 52.288679,-2.075064 52.288727,-2.075535 52.288834,-2.075774 52.288904,-2.076256 52.289002,-2.076744 52.289064,-2.077055 52.289078,-2.078275 52.289072,-2.078724 52.289081,-2.079201 52.289050,-2.079657 52.288989,-2.080043 52.288919,-2.080861 52.288811,-2.081174 52.288777,-2.081460 52.288780,-2.081625 52.288793,-2.081854 52.288827,-2.082291 52.288854,-2.082956 52.288852,-2.083308 52.288815,-2.084300 52.288846,-2.085219 52.288895,-2.085807 52.288953,-2.086436 52.288991,-2.087592 52.289034,-2.088000 52.289104,-2.088219 52.289151,-2.088924 52.289330,-2.089143 52.289416,-2.089221 52.289467,-2.089340 52.289473,-2.089449 52.289454,-2.089574 52.289408,-2.089747 52.289380,-2.089879 52.289347,-2.090188 52.289235,-2.090398 52.289134,-2.090905 52.288765,-2.091065 52.288682,-2.091210 52.288634,-2.091697 52.288546,-2.091690 52.288555,-2.091761 52.288544,-2.091670 52.288663,-2.091704 52.288665,-2.091599 52.288817,-2.091037 52.289604,-2.090062 52.290940,-2.089850 52.291267,-2.089621 52.291537,-2.089496 52.291702,-2.089538 52.291710,-2.089489 52.291750,-2.089442 52.291806,-2.089307 52.291815,-2.088881 52.292379,-2.087762 52.293936,-2.087498 52.294289,-2.086921 52.294974,-2.086501 52.295430,-2.086097 52.295818,-2.085506 52.296304,-2.085518 52.296388,-2.085349 52.296586,-2.084887 52.296939,-2.084896 52.296977,-2.084922 52.296957,-2.085050 52.297304,-2.085253 52.297750,-2.085339 52.297897,-2.085471 52.298248,-2.085610 52.298529,-2.085647 52.298691,-2.086319 52.298609,-2.086429 52.298930,-2.086468 52.298927,-2.086526 52.299067,-2.086638 52.299185,-2.087165 52.300618,-2.087247 52.300612,-2.087466 52.301197,-2.087445 52.301203,-2.087646 52.301548,-2.087754 52.301717,-2.087778 52.301713,-2.087846 52.301842,-2.087453 52.301933,-2.087575 52.302147,-2.087680 52.302359,-2.088059 52.302315,-2.088586 52.303393,-2.088600 52.303445,-2.088639 52.303490,-2.088184 52.303642,-2.088000 52.303716,-2.087934 52.303726,-2.087872 52.303718,-2.087965 52.303825,-2.088306 52.303689,-2.088368 52.303689,-2.088613 52.303651,-2.089002 52.303535,-2.089168 52.303474,-2.089415 52.303412,-2.089627 52.303376,-2.089680 52.303381,-2.089731 52.303398,-2.089808 52.303460,-2.090061 52.303576,-2.090177 52.303638,-2.090240 52.303682,-2.090264 52.303716,-2.090305 52.303817,-2.090353 52.303896,-2.090424 52.304772,-2.090411 52.304924,-2.090368 52.304919,-2.090054 52.305007,-2.089957 52.305056,-2.089914 52.305133,-2.089795 52.305256,-2.089661 52.305412,-2.089570 52.305440,-2.089398 52.305612,-2.089353 52.305671,-2.089157 52.305807,-2.089142 52.305801,-2.089030 52.305938,-2.088884 52.306162,-2.088746 52.306409,-2.088712 52.306486,-2.088691 52.306593,-2.088621 52.306803,-2.088535 52.306999,-2.088504 52.307051,-2.088412 52.307428,-2.088075 52.308255,-2.088001 52.308460,-2.087950 52.308631,-2.087737 52.308893,-2.087422 52.309400,-2.087301 52.309685,-2.087312 52.309752,-2.087305 52.309787,-2.087246 52.309884,-2.087184 52.309943,-2.087100 52.310041,-2.087040 52.310142,-2.086968 52.310340,-2.086917 52.310438,-2.086682 52.310766,-2.086573 52.310978,-2.086508 52.311069,-2.086327 52.311251,-2.086261 52.311301,-2.085715 52.311598,-2.085498 52.311817,-2.085410 52.311928,-2.085356 52.312024,-2.085283 52.312273,-2.085250 52.312691,-2.085369 52.312772,-2.085511 52.312892,-2.085635 52.312967,-2.085903 52.313191,-2.085483 52.313373,-2.085421 52.313383,-2.085171 52.313484,-2.085278 52.313556,-2.085376 52.313642,-2.085635 52.313981,-2.085685 52.314011,-2.085748 52.314078,-2.085976 52.314252,-2.086179 52.314389,-2.086336 52.314511,-2.087168 52.314448,-2.088603 52.314282,-2.090353 52.314114,-2.091637 52.313925,-2.093850 52.313566,-2.095110 52.313352,-2.095063 52.313396,-2.095114 52.313390,-2.095021 52.313455,-2.094980 52.313493,-2.094876 52.313814,-2.094887 52.313851,-2.094860 52.313915,-2.094872 52.313916,-2.094711 52.314213,-2.094600 52.314373,-2.094547 52.314502,-2.094506 52.314562,-2.094396 52.314662,-2.094283 52.314789,-2.094096 52.315121,-2.094117 52.315247,-2.094077 52.315289,-2.094110 52.315364,-2.094138 52.315400,-2.094279 52.315329,-2.095189 52.314764,-2.095501 52.314583,-2.095649 52.314529,-2.096035 52.314452,-2.096696 52.314342,-2.096970 52.314221,-2.097872 52.313866,-2.098594 52.313625,-2.099012 52.313523,-2.099169 52.313505,-2.099232 52.313487,-2.099232 52.313467,-2.099424 52.313476,-2.100191 52.313420,-2.100623 52.313402,-2.101965 52.313385,-2.102478 52.313468,-2.102733 52.313542,-2.103398 52.313667,-2.105261 52.313981,-2.105966 52.314135,-2.106211 52.314167,-2.106413 52.314209,-2.107741 52.314432,-2.108032 52.314465,-2.108610 52.314513,-2.109551 52.314558,-2.110031 52.314569,-2.111369 52.314531,-2.112306 52.314414,-2.113812 52.313707,-2.114038 52.313612,-2.114489 52.313444,-2.114296 52.313163,-2.114314 52.313158,-2.114308 52.313150,-2.115067 52.313092,-2.115593 52.313089,-2.117168 52.313107,-2.117440 52.313136,-2.118038 52.313240,-2.119347 52.313587,-2.119356 52.313577,-2.119987 52.313738,-2.120187 52.313565,-2.120199 52.313568,-2.120218 52.313551,-2.120894 52.313779,-2.121083 52.313832,-2.121394 52.313953,-2.122043 52.314135,-2.122291 52.314189,-2.123042 52.314326,-2.123840 52.314433,-2.124091 52.314483,-2.123785 52.314935,-2.123592 52.315188,-2.123448 52.315320,-2.123084 52.315622,-2.122828 52.315855,-2.122568 52.316112,-2.122506 52.316178,-2.122463 52.316241,-2.122313 52.316491,-2.122203 52.316615,-2.122052 52.316746,-2.121900 52.316853,-2.121795 52.316916,-2.121592 52.317011,-2.121536 52.317025,-2.121396 52.317086,-2.121146 52.317227,-2.120886 52.317362,-2.120798 52.317433,-2.120798 52.317469,-2.120749 52.317496,-2.120662 52.317512,-2.120535 52.317578,-2.120101 52.317836,-2.119744 52.318092,-2.119650 52.318187,-2.119598 52.318265,-2.119505 52.318373,-2.118967 52.318735,-2.118809 52.318860,-2.118619 52.319043,-2.118500 52.319180,-2.118390 52.319273,-2.118112 52.319549,-2.118039 52.319611,-2.117644 52.320018,-2.117261 52.320476,-2.117036 52.320778,-2.116937 52.320929,-2.116827 52.321063,-2.116742 52.321188,-2.116534 52.321428,-2.116464 52.321520,-2.116384 52.321592,-2.116000 52.321856,-2.115902 52.321942,-2.115765 52.322085,-2.116078 52.322157,-2.116434 52.322224,-2.117004 52.322374,-2.117281 52.322427,-2.117917 52.322589,-2.118185 52.322664,-2.118394 52.322734,-2.117899 52.323001,-2.117947 52.323041,-2.118041 52.323157,-2.118106 52.323314,-2.118155 52.323591,-2.118218 52.323777,-2.118240 52.323921,-2.118233 52.324028,-2.118283 52.324157,-2.118291 52.324285,-2.118346 52.324400,-2.118501 52.324440,-2.118493 52.324518,-2.118501 52.324587,-2.118564 52.324878,-2.118597 52.325156,-2.118346 52.325172,-2.117312 52.325162,-2.117316 52.325252,-2.117441 52.325744,-2.117491 52.326015,-2.117604 52.326394,-2.117641 52.326557,-2.117580 52.326613,-2.117480 52.326651,-2.117369 52.326711,-2.117307 52.326770,-2.117239 52.326860,-2.117114 52.327112,-2.117055 52.327253,-2.116620 52.327100,-2.116181 52.327087,-2.116224 52.327300,-2.116294 52.327498,-2.116278 52.327558,-2.116291 52.327653,-2.116331 52.327785,-2.116384 52.327897,-2.116542 52.328152,-2.116882 52.328570,-2.117014 52.328778,-2.117264 52.329278,-2.117316 52.329461,-2.117354 52.329645,-2.117384 52.329733,-2.117488 52.329972,-2.117758 52.330416,-2.118110 52.330938,-2.118206 52.331134,-2.118226 52.331141,-2.118225 52.331165,-2.118358 52.331372,-2.118363 52.331398,-2.118379 52.331397,-2.118375 52.331488,-2.118385 52.331538,-2.118472 52.331748,-2.118489 52.331769,-2.118546 52.331778,-2.118562 52.332037,-2.118548 52.332168,-2.118530 52.332208,-2.118422 52.332358,-2.118330 52.332520,-2.118086 52.332969,-2.117980 52.333191,-2.117965 52.333239,-2.117942 52.333486,-2.117849 52.333487,-2.117661 52.334663,-2.117608 52.335096,-2.117569 52.335274,-2.117546 52.335273,-2.117540 52.335316,-2.117373 52.335171,-2.117240 52.335168,-2.117218 52.335205,-2.117228 52.335228,-2.117228 52.335275,-2.117221 52.335299,-2.117201 52.335321,-2.117154 52.335325,-2.117098 52.335318,-2.117017 52.335328,-2.116976 52.335343,-2.116940 52.335381,-2.116868 52.335416,-2.116830 52.335413,-2.116804 52.335399,-2.116698 52.335407,-2.116693 52.335414,-2.116703 52.335473,-2.116696 52.335492,-2.116664 52.335510,-2.116478 52.335524,-2.116427 52.335519,-2.116249 52.335525,-2.116230 52.335505,-2.116232 52.335489,-2.116110 52.335417,-2.116068 52.335412,-2.116055 52.335402,-2.115825 52.335409,-2.115742 52.335429,-2.115672 52.335421,-2.115665 52.335280,-2.115460 52.335294,-2.115409 52.335285,-2.115186 52.335294,-2.115175 52.335290,-2.115169 52.335271,-2.115181 52.335243,-2.115239 52.335213,-2.115256 52.335194,-2.115236 52.335114,-2.115098 52.335147,-2.114971 52.335258,-2.114736 52.335307,-2.114681 52.335345,-2.114640 52.335391,-2.114588 52.335417,-2.114543 52.335468,-2.114559 52.335485,-2.114614 52.335516,-2.114636 52.335543,-2.114639 52.335576,-2.114615 52.335609,-2.114564 52.335620,-2.114479 52.335593,-2.114428 52.335593,-2.114381 52.335646,-2.114360 52.335692,-2.114278 52.335773,-2.114223 52.335850,-2.114190 52.335861,-2.114120 52.335866,-2.114055 52.335902,-2.114044 52.335938,-2.113947 52.335956,-2.113757 52.336072,-2.113726 52.336142,-2.113689 52.336195,-2.113704 52.336237,-2.113698 52.336268,-2.113685 52.336280,-2.113647 52.336287,-2.113596 52.336280,-2.113521 52.336304,-2.113583 52.336368,-2.113633 52.336388,-2.113644 52.336412,-2.113481 52.336590,-2.113475 52.336669,-2.113491 52.336691,-2.113487 52.336715,-2.113456 52.336739,-2.113390 52.336745,-2.113322 52.336783,-2.113356 52.336805,-2.113431 52.336821,-2.113468 52.336837,-2.113480 52.336849,-2.113480 52.336868,-2.113464 52.336896,-2.113340 52.336901,-2.113275 52.336939,-2.113295 52.336954,-2.113388 52.336970,-2.113404 52.336989,-2.113401 52.337003,-2.113347 52.337045,-2.113290 52.337147,-2.113485 52.337307,-2.113653 52.337381,-2.113746 52.337493,-2.113884 52.337566,-2.113909 52.337594,-2.113941 52.337653,-2.113966 52.337673,-2.114067 52.337718,-2.114145 52.337727,-2.114178 52.337746,-2.114190 52.337832,-2.114212 52.337898,-2.114217 52.337987,-2.114233 52.338019,-2.114222 52.338057,-2.114226 52.338119,-2.114251 52.338195,-2.114291 52.338236,-2.114311 52.338326,-2.114311 52.338398,-2.114303 52.338432,-2.114275 52.338475,-2.114226 52.338511,-2.114293 52.338524,-2.114817 52.338569,-2.115906 52.338688,-2.116485 52.338732,-2.116374 52.338905,-2.116333 52.339069,-2.116337 52.339161,-2.116373 52.339164,-2.116371 52.339209,-2.116426 52.339287,-2.116380 52.339683,-2.116341 52.339791,-2.116303 52.339857,-2.116229 52.339953,-2.116096 52.340166,-2.116780 52.340098,-2.116871 52.340364,-2.116961 52.340546,-2.117126 52.340935,-2.117432 52.341606,-2.117443 52.341691,-2.117415 52.341729,-2.117385 52.341744,-2.117268 52.341752,-2.117037 52.341785,-2.117002 52.341924,-2.117105 52.341972,-2.117333 52.342054,-2.118257 52.342166,-2.118881 52.342171,-2.118937 52.342077,-2.118984 52.342032,-2.119057 52.341923,-2.119173 52.341702,-2.119199 52.341644,-2.119268 52.341429,-2.119387 52.341420,-2.119643 52.341340,-2.119882 52.341209,-2.120199 52.341051,-2.120634 52.340809,-2.121048 52.340624,-2.120996 52.340551,-2.120897 52.340450,-2.121397 52.340281,-2.121708 52.340214,-2.121866 52.340194,-2.123120 52.340106,-2.123918 52.340060,-2.123903 52.340207,-2.124721 52.340227,-2.124719 52.341006,-2.124758 52.341979,-2.123875 52.342144,-2.122818 52.342358,-2.122811 52.342383,-2.122819 52.342411,-2.122897 52.342504,-2.123337 52.342965,-2.123454 52.343075,-2.123610 52.343154,-2.123718 52.343240,-2.123991 52.343389,-2.124168 52.343458,-2.124411 52.343520,-2.124646 52.343606,-2.124852 52.343706,-2.125255 52.343882,-2.125366 52.343948,-2.125388 52.343981,-2.125391 52.344009,-2.125449 52.344109,-2.125655 52.344338,-2.125871 52.344560,-2.126069 52.344858,-2.126167 52.344971,-2.126248 52.345095,-2.126334 52.345199,-2.126366 52.345279,-2.126384 52.345361,-2.126381 52.345506,-2.126389 52.345539,-2.126779 52.345507,-2.127042 52.345475,-2.127828 52.345479,-2.128443 52.345499,-2.128745 52.345543,-2.128808 52.345568,-2.129027 52.345606,-2.129128 52.345656,-2.129192 52.345675,-2.129368 52.345771,-2.130014 52.346279,-2.130254 52.346498,-2.130533 52.346680,-2.130844 52.346852,-2.131172 52.346950,-2.131444 52.347043,-2.131440 52.347055,-2.131477 52.347065,-2.131341 52.347236,-2.131021 52.347695,-2.130887 52.347865,-2.130860 52.347931,-2.130776 52.348005,-2.130291 52.348599,-2.129911 52.349107,-2.129556 52.349620,-2.129539 52.349680,-2.129483 52.349796,-2.129311 52.350279,-2.129222 52.350434,-2.129199 52.350498,-2.129157 52.350548,-2.129092 52.350662,-2.129052 52.350669,-2.129020 52.350716,-2.128995 52.350709,-2.128874 52.350863,-2.128699 52.351036,-2.128599 52.351115,-2.128008 52.351516,-2.127638 52.351729,-2.127536 52.351798,-2.127331 52.351960,-2.127747 52.352230,-2.128179 52.352209,-2.128336 52.352217,-2.128590 52.352213,-2.128747 52.352224,-2.129133 52.352198,-2.129363 52.352099,-2.129477 52.352079,-2.129790 52.351987,-2.130048 52.351787,-2.130735 52.351314,-2.131491 52.350616,-2.131605 52.350491,-2.131622 52.350496,-2.131775 52.350338,-2.131806 52.350284,-2.132012 52.350102,-2.132009 52.350091,-2.132111 52.349876,-2.132377 52.349221,-2.132424 52.349067,-2.132484 52.348928,-2.132765 52.348159,-2.133023 52.347998,-2.133131 52.348034,-2.133156 52.348025,-2.134690 52.348882,-2.135149 52.349076,-2.136057 52.349531,-2.136129 52.349545,-2.136206 52.349579,-2.136264 52.349595,-2.136565 52.349735,-2.137112 52.349942,-2.137173 52.349921,-2.137261 52.349935,-2.137514 52.349922,-2.137672 52.349843,-2.138014 52.349513,-2.138083 52.349432,-2.138199 52.349255,-2.138278 52.349111,-2.138421 52.348883,-2.138595 52.348700,-2.138891 52.348503,-2.139245 52.348351,-2.139602 52.348288,-2.139885 52.348127,-2.140037 52.348100,-2.140074 52.348080,-2.140119 52.348041,-2.140146 52.348031,-2.140204 52.348031,-2.140268 52.348016,-2.140483 52.348020,-2.141926 52.347937,-2.142056 52.347918,-2.142103 52.347669,-2.142113 52.347657,-2.142114 52.347581,-2.142129 52.347542,-2.142187 52.347471,-2.142315 52.347366,-2.142348 52.347324,-2.142421 52.347199,-2.142626 52.346985,-2.142704 52.346914,-2.142816 52.346874,-2.143331 52.346642,-2.143822 52.346436,-2.144303 52.346293,-2.144772 52.346065,-2.145786 52.345503,-2.146264 52.345228,-2.146698 52.345411,-2.146962 52.345505,-2.147218 52.345624,-2.148219 52.346181,-2.148509 52.346380,-2.148608 52.346416,-2.148864 52.346294,-2.148891 52.346270,-2.148974 52.346107,-2.148978 52.346081,-2.149007 52.346025,-2.149030 52.346006,-2.149039 52.345978,-2.149108 52.345940,-2.149160 52.345896,-2.149181 52.345844,-2.149257 52.345777,-2.149298 52.345645,-2.149317 52.345612,-2.149374 52.345575,-2.149548 52.345400,-2.149759 52.345253,-2.149709 52.345200,-2.149683 52.345156,-2.149675 52.345096,-2.149764 52.345011,-2.149760 52.344998,-2.149779 52.344975,-2.149772 52.344967,-2.149701 52.344967,-2.149799 52.344730,-2.149823 52.344689,-2.149891 52.344633,-2.149984 52.344584,-2.150003 52.344540,-2.150047 52.344485,-2.150144 52.344423,-2.150375 52.344195,-2.150561 52.344025,-2.150953 52.343686,-2.151118 52.343234,-2.151170 52.342818,-2.151196 52.342705,-2.151230 52.342610,-2.151266 52.342542,-2.151337 52.342437,-2.151588 52.342507,-2.151707 52.342175,-2.151792 52.341803,-2.151815 52.341739,-2.151839 52.341705,-2.151912 52.341632,-2.151961 52.341521,-2.151947 52.341479,-2.151954 52.341336,-2.151931 52.341175,-2.151925 52.341055,-2.151935 52.341010,-2.152010 52.340820,-2.151978 52.340706,-2.151973 52.340636,-2.151990 52.340528,-2.151986 52.340446,-2.152005 52.340384,-2.152052 52.340328,-2.152141 52.340091,-2.152411 52.339611,-2.152532 52.339230,-2.152599 52.338781,-2.152571 52.338687,-2.152361 52.338695,-2.152306 52.338704,-2.152303 52.338697,-2.152141 52.338125,-2.152111 52.337983,-2.152015 52.337841,-2.152023 52.337755,-2.151975 52.337663,-2.151921 52.337607,-2.151889 52.337587,-2.151896 52.337568,-2.151882 52.337481,-2.151830 52.337423,-2.151824 52.337399,-2.151841 52.337305,-2.151995 52.337192,-2.151927 52.337036,-2.151925 52.336990,-2.151919 52.336985,-2.152022 52.336898,-2.152214 52.336713,-2.152252 52.336693,-2.152416 52.336642,-2.152624 52.336476,-2.152681 52.336407,-2.152807 52.336293,-2.152850 52.336239,-2.152925 52.336192,-2.153408 52.335725,-2.153758 52.335705,-2.153848 52.335684,-2.153882 52.335665,-2.154056 52.335496,-2.154262 52.335348,-2.154398 52.335207,-2.154458 52.335126,-2.154538 52.335043,-2.154581 52.334941,-2.154721 52.334768,-2.154731 52.334667,-2.154844 52.334455,-2.154820 52.334451,-2.154831 52.334421,-2.154820 52.334377,-2.154787 52.334351,-2.154813 52.334327,-2.154816 52.334309,-2.154831 52.334310,-2.154856 52.334279,-2.154961 52.334211,-2.155008 52.334142,-2.155018 52.334117,-2.155022 52.334047,-2.154978 52.333982,-2.154989 52.333945,-2.155010 52.333923,-2.155272 52.333816,-2.155285 52.333801,-2.155303 52.333735,-2.155366 52.333674,-2.155509 52.333581,-2.155527 52.333581,-2.155645 52.333636,-2.155684 52.333644,-2.155755 52.333640,-2.155830 52.333625,-2.155951 52.333567,-2.156158 52.333440,-2.156200 52.333390,-2.156225 52.333349,-2.156251 52.333274,-2.156250 52.333219,-2.156205 52.333147,-2.156215 52.333124,-2.156364 52.332974,-2.156488 52.332837,-2.156512 52.332801,-2.156415 52.332365,-2.156256 52.331969,-2.156245 52.331759,-2.156167 52.331527,-2.156074 52.331323,-2.156055 52.331098,-2.156072 52.331021,-2.156137 52.330972,-2.156224 52.331001,-2.156520 52.331045,-2.156667 52.331058,-2.156922 52.331058,-2.157150 52.331034,-2.157615 52.330952,-2.158323 52.330782,-2.159536 52.330378,-2.159914 52.330213,-2.160154 52.329931,-2.160362 52.329721,-2.160482 52.329617,-2.160716 52.329441,-2.161531 52.328780,-2.161775 52.328570,-2.161811 52.328500,-2.161991 52.328437,-2.162078 52.328382,-2.162136 52.328330,-2.162230 52.328273,-2.162497 52.328146,-2.162832 52.327858,-2.163012 52.327683,-2.163238 52.327495,-2.163504 52.327221,-2.163642 52.327119,-2.163685 52.327063,-2.163755 52.326934,-2.163814 52.326878,-2.163904 52.326818,-2.163969 52.326719,-2.164095 52.326580,-2.164155 52.326499,-2.164205 52.326405,-2.164244 52.326308,-2.164321 52.326195,-2.164351 52.326137,-2.164361 52.326091,-2.164382 52.326064,-2.164382 52.326039,-2.164340 52.326017,-2.164461 52.325940,-2.164749 52.325733,-2.165003 52.325616,-2.165200 52.325552,-2.165764 52.325421,-2.165973 52.325348,-2.165961 52.325286,-2.166443 52.325076,-2.166552 52.325007,-2.166641 52.324964,-2.166930 52.324774,-2.167050 52.324727,-2.167042 52.324712,-2.167168 52.324656,-2.167251 52.324630,-2.167556 52.324508,-2.167583 52.324483,-2.167617 52.324433,-2.167637 52.324384,-2.167635 52.324333,-2.167643 52.324309,-2.167688 52.324207,-2.167747 52.324111,-2.167768 52.323804,-2.167766 52.323644,-2.167805 52.323279,-2.167736 52.323146,-2.167752 52.322988,-2.167681 52.322773,-2.167652 52.322712,-2.167619 52.322562,-2.167602 52.322261,-2.167597 52.321704,-2.167563 52.321542,-2.167577 52.321386,-2.167565 52.321306,-2.167577 52.321226,-2.167554 52.321062,-2.167527 52.320962,-2.167557 52.320904,-2.167550 52.320719,-2.167564 52.320559,-2.167516 52.320391,-2.167507 52.320309,-2.167524 52.320227,-2.167551 52.320153,-2.167613 52.320029,-2.167707 52.319887,-2.167871 52.319691,-2.167837 52.319551,-2.167869 52.319431,-2.167870 52.319341,-2.167855 52.319293,-2.167868 52.319207,-2.167815 52.319076,-2.167829 52.318983,-2.167820 52.318983,-2.167827 52.318943,-2.167818 52.318923,-2.167807 52.318911,-2.167724 52.318873,-2.167593 52.318830,-2.167565 52.318812,-2.167534 52.318770,-2.167461 52.318567,-2.167190 52.318458,-2.167019 52.318429,-2.166805 52.318374,-2.166845 52.318353,-2.166812 52.318346,-2.166816 52.318337,-2.166873 52.318280,-2.167197 52.318118,-2.167337 52.318059,-2.167626 52.317981,-2.167828 52.317841,-2.167966 52.317805,-2.168157 52.317720,-2.168325 52.317613,-2.168364 52.317579,-2.168428 52.317485,-2.168522 52.317403,-2.168578 52.317366,-2.168703 52.317298,-2.168802 52.317259,-2.169007 52.317203,-2.169080 52.317187,-2.169202 52.317176,-2.169368 52.317183,-2.169461 52.317170,-2.169515 52.317156,-2.169651 52.317066,-2.169737 52.317055,-2.170013 52.317096,-2.170343 52.317102,-2.170484 52.317114,-2.170837 52.317174,-2.170985 52.317190,-2.171361 52.317270,-2.171475 52.317279,-2.171796 52.317280,-2.172001 52.317267,-2.172172 52.317218,-2.172637 52.317211,-2.172726 52.317221,-2.172919 52.317223,-2.173170 52.317205,-2.173231 52.317191,-2.173338 52.317090,-2.173571 52.316962,-2.173787 52.316803,-2.173968 52.316718,-2.174129 52.316665,-2.174429 52.316603,-2.174585 52.316600,-2.174595 52.316639,-2.174623 52.316640,-2.174671 52.316820,-2.174690 52.316962,-2.174903 52.317702,-2.174976 52.317861,-2.175098 52.318394,-2.175111 52.318512,-2.175105 52.318723,-2.175137 52.318864,-2.175171 52.318955,-2.175217 52.319173,-2.175312 52.319823,-2.175281 52.320017,-2.175092 52.320602,-2.174953 52.320846,-2.175157 52.320862,-2.175128 52.320925,-2.175161 52.320934,-2.175087 52.321118,-2.175078 52.321188,-2.175088 52.321285,-2.175086 52.321344,-2.175104 52.321455,-2.175203 52.321838,-2.175208 52.322077,-2.175234 52.322194,-2.175235 52.322248,-2.175192 52.322461,-2.175148 52.322798,-2.175018 52.323163,-2.174992 52.323374,-2.174922 52.323605,-2.174831 52.323865,-2.174762 52.324124,-2.174716 52.324379,-2.174706 52.324609,-2.174627 52.324921,-2.174568 52.325072,-2.174475 52.325273,-2.174345 52.325650,-2.174240 52.325909,-2.174163 52.326047,-2.174062 52.326181,-2.174041 52.326226,-2.174034 52.326264,-2.173781 52.326794,-2.173714 52.326955,-2.173621 52.327099,-2.173440 52.327322,-2.173443 52.327399,-2.173305 52.327594,-2.173265 52.327665,-2.173226 52.327711,-2.173188 52.327735,-2.173136 52.327797,-2.173132 52.327808,-2.173160 52.327839,-2.173060 52.327966,-2.172849 52.328103,-2.172617 52.328220,-2.172411 52.328262,-2.172314 52.328291,-2.172239 52.328334,-2.172190 52.328381,-2.172082 52.328536,-2.172051 52.328651,-2.172027 52.328704,-2.171989 52.328767,-2.171897 52.328879,-2.171729 52.329352,-2.171692 52.329480,-2.171620 52.329608,-2.171626 52.329610,-2.171556 52.329714,-2.171551 52.329734,-2.171225 52.330148,-2.171069 52.330395,-2.170947 52.330645,-2.170744 52.331426,-2.170772 52.331516,-2.170925 52.331642,-2.171090 52.331704,-2.171202 52.331718,-2.171623 52.331686,-2.172261 52.331684,-2.172524 52.331692,-2.172665 52.331706,-2.172735 52.331707,-2.172817 52.331696,-2.172994 52.331713,-2.173220 52.331718,-2.173343 52.331713,-2.173420 52.331723,-2.173660 52.331728,-2.173927 52.331666,-2.174040 52.331648,-2.174143 52.331452,-2.174194 52.331397,-2.174286 52.331356,-2.174374 52.331343,-2.175252 52.331355,-2.175561 52.331392,-2.175703 52.331399,-2.175772 52.331424,-2.175827 52.331471,-2.176049 52.331552,-2.176318 52.331621,-2.176565 52.331660,-2.176894 52.331696,-2.177422 52.331700,-2.177705 52.331757,-2.177882 52.331756,-2.178003 52.331702,-2.178084 52.331638,-2.178135 52.331583,-2.178189 52.331555,-2.178264 52.331495,-2.178471 52.331432,-2.178580 52.331417,-2.178718 52.331377,-2.178949 52.331337,-2.179223 52.331255,-2.179737 52.331181,-2.180123 52.331140,-2.180877 52.331100,-2.181001 52.331085,-2.181238 52.331034,-2.181453 52.330966,-2.181606 52.330892,-2.182079 52.330633,-2.182259 52.330575,-2.182558 52.330432,-2.182666 52.330371,-2.182889 52.330308,-2.182955 52.330296,-2.183352 52.330261,-2.183495 52.330224,-2.183557 52.330171,-2.183584 52.330159,-2.184212 52.330135,-2.184803 52.330100,-2.184958 52.330076,-2.185085 52.330038,-2.185572 52.329844,-2.185688 52.329784,-2.185824 52.329743,-2.186060 52.329691,-2.186245 52.329610,-2.186349 52.329575,-2.186471 52.329541,-2.186729 52.329489,-2.187065 52.329331,-2.187199 52.329225,-2.187320 52.329180,-2.187527 52.329169,-2.187637 52.329201,-2.187696 52.329209,-2.188117 52.329218,-2.188283 52.329183,-2.188438 52.329176,-2.189051 52.329076,-2.189045 52.329056,-2.189104 52.329048,-2.189338 52.329087,-2.189494 52.329083,-2.189647 52.329019,-2.189885 52.328899,-2.190011 52.328871,-2.190265 52.328875,-2.190386 52.328899,-2.190497 52.328932,-2.190507 52.329013,-2.190498 52.329242,-2.190519 52.329528,-2.190581 52.329874,-2.190581 52.329917,-2.190553 52.330558,-2.190469 52.330901,-2.190495 52.331033,-2.190495 52.331106,-2.190352 52.331540,-2.190358 52.331598,-2.190444 52.331875,-2.190480 52.331950,-2.190545 52.332041,-2.190718 52.332205,-2.190914 52.332367,-2.191054 52.332530,-2.191102 52.332575,-2.191222 52.332684,-2.191465 52.332881,-2.191563 52.333012,-2.191623 52.333049,-2.191964 52.333160,-2.192041 52.333194,-2.192186 52.333301,-2.192457 52.333450,-2.192489 52.333476,-2.192519 52.333510,-2.192538 52.333550,-2.192577 52.333665,-2.192650 52.333936,-2.192656 52.333982,-2.192629 52.334140,-2.192633 52.334196,-2.192594 52.334518,-2.192606 52.334542,-2.192672 52.334617,-2.192913 52.334780,-2.192980 52.334800,-2.193103 52.334792,-2.193148 52.334798,-2.193198 52.334821,-2.193209 52.334838,-2.193197 52.334894,-2.193146 52.334988,-2.193161 52.335029,-2.193232 52.335111,-2.193332 52.335182,-2.193382 52.335231,-2.193480 52.335262,-2.193494 52.335283,-2.193492 52.335308,-2.193485 52.335314,-2.193375 52.335350,-2.193394 52.335391,-2.193481 52.335475,-2.193496 52.335502,-2.193490 52.335517,-2.193438 52.335561,-2.193321 52.335581,-2.193233 52.335622,-2.193227 52.335645,-2.193245 52.335709,-2.193240 52.335770,-2.193258 52.335842,-2.193197 52.336003,-2.193185 52.336087,-2.193176 52.336102,-2.193133 52.336109,-2.192998 52.336066,-2.192849 52.335968,-2.192791 52.335988,-2.192762 52.336014,-2.192753 52.336049,-2.192763 52.336073,-2.192833 52.336148,-2.192827 52.336163,-2.192758 52.336242,-2.192590 52.336367,-2.192520 52.336457,-2.192513 52.336479,-2.192519 52.336494,-2.192610 52.336561,-2.192661 52.336610,-2.192739 52.336623,-2.192810 52.336607,-2.192871 52.336610,-2.192877 52.336625,-2.192866 52.336642,-2.192740 52.336681,-2.192709 52.336711,-2.192727 52.336760,-2.192809 52.336855,-2.192832 52.336891,-2.192832 52.336910,-2.192821 52.336918,-2.192780 52.336921,-2.192772 52.336931,-2.192492 52.336888,-2.192412 52.336950,-2.192409 52.336973,-2.192212 52.336915,-2.192173 52.336927,-2.192116 52.337023,-2.192100 52.337094,-2.192034 52.337152,-2.192005 52.337191,-2.192008 52.337208,-2.192023 52.337221,-2.192082 52.337246,-2.192107 52.337270,-2.192104 52.337281,-2.192031 52.337349,-2.192023 52.337366,-2.192019 52.337412,-2.192040 52.337524,-2.192052 52.337543,-2.192075 52.337556,-2.192341 52.337458,-2.192448 52.337476,-2.192485 52.337496,-2.192514 52.337529,-2.192535 52.337541,-2.192539 52.337555,-2.192529 52.337621,-2.192518 52.337639,-2.192477 52.337671,-2.192399 52.337708,-2.192383 52.337726,-2.192398 52.337765,-2.192472 52.337843,-2.192574 52.337809,-2.192678 52.337817,-2.192722 52.337810,-2.192874 52.337903,-2.192996 52.337918,-2.193027 52.337930,-2.193093 52.338026,-2.193088 52.338068,-2.193058 52.338091,-2.192932 52.338135,-2.192915 52.338158,-2.192921 52.338181,-2.192996 52.338232,-2.193039 52.338286,-2.193042 52.338301,-2.192993 52.338516,-2.193118 52.338536,-2.193241 52.338540,-2.193279 52.338558,-2.193298 52.338583,-2.193282 52.338610,-2.193197 52.338662,-2.193114 52.338732,-2.193091 52.338780,-2.193093 52.338866,-2.193113 52.338890,-2.193146 52.338896,-2.193564 52.338869,-2.193659 52.338905,-2.193684 52.338926,-2.193687 52.338945,-2.193680 52.338959,-2.193649 52.338977,-2.193463 52.338987,-2.193412 52.339027,-2.193402 52.339068,-2.193431 52.339104,-2.193590 52.339167,-2.193627 52.339186,-2.193647 52.339209,-2.193649 52.339226,-2.193563 52.339374,-2.193567 52.339389,-2.193598 52.339433,-2.193759 52.339536,-2.193777 52.339573,-2.193777 52.339598,-2.193830 52.339621,-2.193902 52.339683,-2.193899 52.339871,-2.193917 52.339934,-2.193949 52.339997,-2.194064 52.340045,-2.194340 52.340085,-2.194694 52.340170,-2.195063 52.340284,-2.195192 52.340342,-2.195538 52.340588,-2.195613 52.340654,-2.195651 52.340922,-2.195722 52.341019,-2.195756 52.341129,-2.195596 52.341258,-2.195703 52.341304,-2.196059 52.341489,-2.196331 52.341650,-2.196752 52.341869,-2.196998 52.342017,-2.197936 52.342553,-2.198575 52.342886,-2.198828 52.343041,-2.199158 52.343221,-2.200485 52.343984,-2.201049 52.344281,-2.201745 52.344605,-2.202135 52.344849,-2.202169 52.344856,-2.202238 52.344809,-2.202638 52.344986,-2.202921 52.345142,-2.203049 52.345178,-2.203315 52.345219,-2.203513 52.345199,-2.203597 52.345217,-2.203644 52.345179,-2.204124 52.345269,-2.204309 52.345315,-2.204569 52.345398,-2.205080 52.345612,-2.205099 52.345638,-2.205105 52.345683,-2.205088 52.345707,-2.205030 52.345754,-2.205020 52.345788,-2.205017 52.345892,-2.204873 52.346254,-2.204838 52.346389,-2.204810 52.346723,-2.204777 52.346866,-2.204728 52.347009,-2.204743 52.347142,-2.204744 52.347313,-2.204754 52.347440,-2.204727 52.347721,-2.204733 52.347763,-2.204728 52.348025,-2.204778 52.348823,-2.204796 52.348951,-2.204833 52.349110,-2.204961 52.349439,-2.205051 52.349579,-2.205256 52.349815,-2.205339 52.349925,-2.205385 52.350014,-2.205412 52.350099,-2.205434 52.350334,-2.205419 52.350407,-2.205373 52.350555,-2.205348 52.350594,-2.205321 52.350703,-2.205346 52.350721,-2.205909 52.350915,-2.206659 52.351241,-2.207171 52.351429,-2.207208 52.351407,-2.207349 52.351465,-2.207417 52.351501,-2.207661 52.351690,-2.207641 52.351705,-2.207629 52.351739,-2.207621 52.351829,-2.207586 52.351950,-2.207672 52.352092,-2.207715 52.352183,-2.207741 52.352279,-2.207849 52.352506,-2.207908 52.352587,-2.208043 52.352724,-2.208069 52.352766,-2.207994 52.352947,-2.207894 52.353235,-2.208065 52.353290,-2.208416 52.353464,-2.208818 52.353708,-2.209078 52.353852,-2.210227 52.354266,-2.210374 52.354342,-2.210403 52.354321,-2.210623 52.354425,-2.210486 52.354585,-2.210212 52.354876,-2.210192 52.354925,-2.210201 52.354952,-2.210358 52.355064,-2.210416 52.355088,-2.210535 52.355170,-2.210669 52.355230,-2.210857 52.355344,-2.211305 52.355563,-2.211635 52.355742,-2.211844 52.355898,-2.212201 52.356073,-2.212513 52.356210,-2.212843 52.356341,-2.213949 52.356835,-2.213975 52.356844,-2.214008 52.356825,-2.214311 52.356936,-2.215069 52.357296,-2.216381 52.357974,-2.217504 52.358453,-2.218162 52.358717,-2.219002 52.359069,-2.219219 52.359138,-2.220211 52.359378,-2.220471 52.359422,-2.221685 52.359671,-2.223059 52.359918,-2.223321 52.359918,-2.224476 52.360181,-2.224934 52.360309,-2.225276 52.360421,-2.225417 52.360449,-2.226117 52.360489,-2.226391 52.360522,-2.230048 52.361069,-2.230380 52.361108,-2.230882 52.361148,-2.231879 52.361187,-2.232138 52.361205,-2.233110 52.361126,-2.233739 52.361018,-2.235523 52.360881),(-2.231313 52.330669,-2.231657 52.330381,-2.231729 52.330310,-2.232135 52.329733,-2.232164 52.329681,-2.232322 52.329311,-2.232310 52.329217,-2.232267 52.329073,-2.232237 52.329015,-2.231435 52.328961,-2.231458 52.328493,-2.232054 52.328423,-2.232023 52.328320,-2.231973 52.328052,-2.231954 52.327913,-2.231939 52.327639,-2.231899 52.327566,-2.231832 52.326595,-2.231818 52.326589,-2.231808 52.326473,-2.231802 52.326227,-2.231808 52.326105,-2.232088 52.326066,-2.232677 52.326051,-2.232668 52.326620,-2.232656 52.326895,-2.233820 52.326880,-2.234332 52.327213,-2.235020 52.327550,-2.235316 52.327703,-2.235853 52.327810,-2.236667 52.327998,-2.237560 52.328242,-2.237491 52.328316,-2.237531 52.328369,-2.237248 52.328628,-2.237060 52.328822,-2.236899 52.329048,-2.235375 52.330839,-2.234811 52.331537,-2.234778 52.331540,-2.234783 52.331588,-2.235739 52.331666,-2.235728 52.331899,-2.235948 52.331908,-2.235949 52.331966,-2.236180 52.332005,-2.236224 52.331986,-2.236311 52.331919,-2.236357 52.331908,-2.236415 52.331903,-2.236572 52.331913,-2.236515 52.332023,-2.236433 52.332244,-2.236399 52.332387,-2.236384 52.332510,-2.236645 52.332567,-2.236622 52.332614,-2.236542 52.332722,-2.236501 52.332749,-2.236216 52.333127,-2.236017 52.333370,-2.235955 52.333468,-2.236222 52.333545,-2.235848 52.334071,-2.235810 52.334386,-2.235953 52.334372,-2.235981 52.334463,-2.236074 52.334488,-2.236107 52.334655,-2.236222 52.334657,-2.236232 52.334691,-2.236980 52.334662,-2.236809 52.335090,-2.236595 52.335514,-2.237030 52.335665,-2.236894 52.335754,-2.237107 52.335855,-2.236895 52.336012,-2.236663 52.336149,-2.236598 52.336210,-2.236770 52.336227,-2.236972 52.336268,-2.237194 52.336347,-2.237479 52.336480,-2.237425 52.336553,-2.237417 52.336636,-2.237403 52.336670,-2.237310 52.336806,-2.237095 52.337033,-2.237035 52.337013,-2.236893 52.337186,-2.236888 52.337268,-2.236822 52.337347,-2.236712 52.337538,-2.236657 52.337586,-2.236609 52.337589,-2.236503 52.337564,-2.236507 52.337541,-2.236476 52.337504,-2.235825 52.336926,-2.235453 52.336844,-2.235411 52.336882,-2.235365 52.336896,-2.235288 52.336898,-2.235195 52.336889,-2.235036 52.336861,-2.234962 52.336840,-2.234898 52.336811,-2.234796 52.336748,-2.234752 52.336709,-2.234647 52.336593,-2.234469 52.336362,-2.234380 52.336259,-2.234312 52.336192,-2.234271 52.336165,-2.234146 52.336115,-2.233681 52.336004,-2.233345 52.335865,-2.233239 52.335832,-2.233153 52.335817,-2.233175 52.335778,-2.233090 52.335749,-2.233044 52.335766,-2.232913 52.335718,-2.232751 52.335583,-2.232668 52.335498,-2.232294 52.335289,-2.232200 52.335215,-2.232050 52.335140,-2.231695 52.335008,-2.231672 52.335040,-2.231469 52.334969,-2.231205 52.334899,-2.230951 52.335169,-2.230825 52.335349,-2.230803 52.335417,-2.230775 52.335453,-2.230493 52.335693,-2.230409 52.335774,-2.230230 52.335906,-2.229941 52.336199,-2.229736 52.336126,-2.229472 52.336045,-2.229283 52.336005,-2.229111 52.335979,-2.228468 52.335927,-2.227869 52.335895,-2.227471 52.335854,-2.227465 52.335808,-2.227496 52.335643,-2.227574 52.335405,-2.227680 52.335031,-2.227711 52.334799,-2.227628 52.334795,-2.227243 52.334880,-2.227093 52.334882,-2.226993 52.334852,-2.226841 52.334749,-2.226608 52.334628,-2.226664 52.334481,-2.226779 52.334495,-2.226762 52.334482,-2.226799 52.334405,-2.226833 52.334393,-2.226928 52.334163,-2.226969 52.334083,-2.227039 52.333980,-2.227670 52.333978,-2.227786 52.333362,-2.227841 52.333226,-2.227940 52.333054,-2.227989 52.332986,-2.228252 52.332731,-2.229113 52.332081,-2.229285 52.331965,-2.229355 52.331924,-2.229514 52.331952,-2.230034 52.331623,-2.230079 52.331523,-2.230347 52.331313,-2.231313 52.330669),(-2.179022 52.310396,-2.179891 52.310464,-2.180426 52.310460,-2.180616 52.310473,-2.180768 52.310497,-2.180773 52.310529,-2.180688 52.310767,-2.180648 52.310848,-2.180618 52.310989,-2.180566 52.311085,-2.180518 52.311210,-2.180429 52.311395,-2.180434 52.311410,-2.181208 52.311829,-2.181117 52.311886,-2.181047 52.311916,-2.180916 52.312004,-2.180683 52.312186,-2.180416 52.312426,-2.180273 52.312569,-2.180068 52.312795,-2.179758 52.313213,-2.179442 52.313514,-2.179172 52.313748,-2.178825 52.314023,-2.178976 52.314065,-2.179431 52.314149,-2.179788 52.314225,-2.179558 52.314728,-2.179657 52.314743,-2.179666 52.314809,-2.179509 52.315053,-2.179403 52.315272,-2.179487 52.315734,-2.179438 52.315846,-2.179480 52.315922,-2.179480 52.315977,-2.179514 52.316186,-2.179523 52.316377,-2.179494 52.316420,-2.179267 52.316551,-2.179197 52.316578,-2.178861 52.316542,-2.177617 52.316485,-2.177489 52.316464,-2.177335 52.316406,-2.177065 52.316278,-2.177110 52.315985,-2.177195 52.315584,-2.177379 52.315002,-2.177119 52.314940,-2.177052 52.314944,-2.176890 52.314930,-2.176608 52.314885,-2.176318 52.314825,-2.175729 52.314775,-2.175381 52.314719,-2.175182 52.314709,-2.175251 52.314403,-2.175798 52.314440,-2.175839 52.314349,-2.176204 52.314412,-2.176739 52.314487,-2.176760 52.314411,-2.176779 52.314179,-2.176580 52.314148,-2.176621 52.314077,-2.176621 52.314010,-2.176596 52.313832,-2.176882 52.313844,-2.177292 52.313811,-2.177248 52.313563,-2.177239 52.313396,-2.177257 52.313137,-2.177274 52.313034,-2.177310 52.312897,-2.177328 52.312846,-2.177340 52.312837,-2.177366 52.312744,-2.177605 52.312243,-2.177592 52.312206,-2.177617 52.312145,-2.177699 52.311988,-2.177904 52.311540,-2.177999 52.311303,-2.178016 52.311282,-2.178044 52.311271,-2.178070 52.311108,-2.178221 52.310726,-2.178243 52.310690,-2.178428 52.310117,-2.178445 52.309985,-2.178418 52.309931,-2.178419 52.309719,-2.178396 52.309566,-2.178368 52.309472,-2.178276 52.309294,-2.179224 52.309159,-2.179246 52.309399,-2.179258 52.309658,-2.179237 52.309768,-2.179213 52.309816,-2.179200 52.309976,-2.179165 52.310094,-2.179084 52.310184,-2.179022 52.310396),(-2.119588 52.294126,-2.119156 52.294717,-2.118880 52.294998,-2.118651 52.295207,-2.118479 52.295147,-2.118220 52.295079,-2.118144 52.295081,-2.118093 52.295096,-2.118001 52.295167,-2.117819 52.295079,-2.117694 52.294985,-2.117564 52.294866,-2.117320 52.294614,-2.117154 52.294467,-2.116923 52.294202,-2.116576 52.293870,-2.115397 52.292876,-2.115299 52.292939,-2.115354 52.292989,-2.114965 52.293248,-2.114625 52.293051,-2.114526 52.293092,-2.114357 52.293205,-2.113672 52.293703,-2.113582 52.293832,-2.113345 52.294000,-2.113240 52.294103,-2.113055 52.294314,-2.112694 52.294206,-2.113142 52.293841,-2.113204 52.293782,-2.112930 52.293676,-2.112789 52.293609,-2.112455 52.293469,-2.112367 52.293448,-2.112255 52.293404,-2.112081 52.293324,-2.111216 52.292859,-2.110677 52.292517,-2.109852 52.292111,-2.109790 52.292062,-2.109674 52.291991,-2.109091 52.291728,-2.108843 52.291629,-2.108674 52.291574,-2.108500 52.291485,-2.107887 52.291202,-2.107546 52.291070,-2.107304 52.290998,-2.107161 52.290966,-2.106825 52.290924,-2.106447 52.290915,-2.106122 52.290895,-2.105633 52.290837,-2.105667 52.290690,-2.105664 52.290592,-2.105717 52.290494,-2.105798 52.290413,-2.105916 52.290376,-2.106170 52.290273,-2.106526 52.290081,-2.106783 52.289967,-2.107428 52.289556,-2.107703 52.289405,-2.108111 52.289155,-2.108360 52.289018,-2.108549 52.288932,-2.108689 52.288859,-2.108749 52.288818,-2.109105 52.288640,-2.109436 52.288450,-2.109756 52.288248,-2.109926 52.288163,-2.110305 52.287948,-2.110537 52.287829,-2.110759 52.287729,-2.111240 52.287546,-2.111354 52.287491,-2.111522 52.287432,-2.112074 52.287210,-2.112430 52.287044,-2.113090 52.286680,-2.113270 52.286602,-2.113539 52.286509,-2.113903 52.286439,-2.114339 52.286368,-2.114722 52.286346,-2.115120 52.286367,-2.115435 52.286399,-2.115567 52.286426,-2.115797 52.286491,-2.115909 52.286547,-2.116039 52.286670,-2.116077 52.286733,-2.116109 52.286827,-2.116136 52.286864,-2.116164 52.286946,-2.116186 52.286979,-2.116289 52.287201,-2.116420 52.287336,-2.116505 52.287398,-2.116630 52.287470,-2.116847 52.287556,-2.117023 52.287593,-2.117335 52.287640,-2.117442 52.287673,-2.117467 52.287700,-2.117530 52.287810,-2.117597 52.287896,-2.117855 52.288115,-2.118489 52.288625,-2.118613 52.288748,-2.119051 52.288733,-2.119252 52.288705,-2.119446 52.288667,-2.119570 52.288689,-2.119711 52.288737,-2.119817 52.288791,-2.119896 52.288845,-2.120117 52.289131,-2.120188 52.289282,-2.120259 52.289465,-2.120410 52.289965,-2.120580 52.290662,-2.120615 52.290873,-2.120642 52.291148,-2.120635 52.291613,-2.120591 52.291826,-2.120576 52.291822,-2.120518 52.292196,-2.120445 52.292538,-2.120191 52.293109,-2.120018 52.293460,-2.119973 52.293531,-2.119588 52.294126)))"  # noqa: E501
-    normalised = wkt.normalise(value, issues=issues)
-    assert normalised == expected
-
-
-def test_wkt_invalid_geometry_self_intersection_normalised():
-    from tests.data.wkt import self_intersections_osgb_wkt
-
-    wkt = WktDataType()
-    issues = IssueLog()
-
-    for wkt_geom in self_intersections_osgb_wkt:
-        osgb_wkt = wkt.normalise(wkt_geom, issues=issues)
-        assert osgb_wkt is not None
-
-        geometry = shapely.wkt.loads(osgb_wkt)
-        assert geometry.is_valid
-
-
-def test_wkt_invalid_multipolygon_too_few_points_normalised():
-    from tests.data.wkt import too_few_points_osgb_wkt
-
-    wkt = WktDataType()
-    issues = IssueLog()
-
-    for wkt_geom in too_few_points_osgb_wkt:
-        osgb_wkt = wkt.normalise(wkt_geom, issues=issues)
-        assert osgb_wkt is not None
-
-        geometry = shapely.wkt.loads(osgb_wkt)
-        assert geometry.is_valid
